### 2020-07-03 494f1fd
Update paper.md
