### 2020-07-03 8f8a02c
Update paper.md
