+++ paper/paper.md
@@ -32,17 +32,8 @@ perform custom calculations, and make graphics.
 
 The workflow of ``ocean_data_tools`` is to build uniform structs (e.g. ``argo``,
 ``cruise``, ``hycom``, ``mercator``, ``woa``, ``wod``) from raw datasets and 
-call general functions on these structs to map, subset, or plot. 
-
-![Building a virtual cruise from the Operational Mercator global ocean
-analysis and forecast system at 1/12 degree with 3D bathymetry [@Smith:1997]. 
-Showing (a) a 3D velocity plot created using ``model_domain_plot``, (b) 
-virtual cruise selection using ``transect_select``, and ``model_build_profiles``, 
-(c) coordinates of the resulting uniform struct, and (d) a temperature section 
-plotted using ``general_section`` with ``bathymetry_section``. Three of the 
-subplots use colormaps from cmocean [@Thyng:2016]. \label{fig:1}](figure.png)
-
-Functions with the ``\_build`` suffix load raw data into uniform structs. Structs are 
+call general functions on these structs to map, subset, or plot. Functions with 
+the ``\_build`` suffix load raw data into uniform structs. Structs are 
 compatible with all ``general_`` functions. One application of the ``\_build`` 
 feature is to create virtual cruises from model output \autoref{fig:1}. The user
 draws transects on a map (or passes coordinates as an argument) to build casts 
@@ -56,6 +47,14 @@ is to continuously add support for additional datasets such as Remote Sensing
 Systems (RSS) and European Centre for Medium-Range 
 Weather Forecasts (ECMWF) products. The source code for ``ocean_data_tools`` has
 been archived to Zenodo with the linked DOI: [@Ferris:2020].
+
+![Building a virtual cruise from the Operational Mercator global ocean
+analysis and forecast system at 1/12 degree with 3D bathymetry [@Smith:1997]. 
+Showing (a) a 3D velocity plot created using ``model_domain_plot``, (b) 
+virtual cruise selection using ``transect_select``, and ``model_build_profiles``, 
+(c) coordinates of the resulting uniform struct, and (d) a temperature section 
+plotted using ``general_section`` with ``bathymetry_section``. Three of the 
+subplots use colormaps from cmocean [@Thyng:2016]. \label{fig:1}](figure.png)
 
 # Acknowledgements
 
