+++ paper/paper.md
@@ -14,7 +14,7 @@ date: 03 July 2020
 bibliography: paper.bib
 ---
 
-# Statement of Need
+# Summary
 
 ``ocean_data_tools`` simplifies the process of extracting, formatting, and 
 visualizing freely-available oceanographic data. A wealth of oceanographic 
@@ -28,7 +28,13 @@ solves this problem by allowing the user to transform common oceanographic
 data sources into uniform structs, call general functions on these structs, 
 perform custom calculations, and make graphics.
 
-# Summary
+![Building a virtual cruise from the Operational Mercator global ocean
+analysis and forecast system at 1/12 degree with 3D bathymetry [@Smith:1997]. 
+Showing (a) a 3D velocity plot created using ``model_domain_plot``, (b) 
+virtual cruise selection using ``transect_select``, and ``model_build_profiles``, 
+(c) coordinates of the resulting uniform struct, and (d) a temperature section 
+plotted using ``general_section`` with ``bathymetry_section``. Three of the 
+subplots use colormaps from cmocean [@Thyng:2016]. \label{fig:1}](figure.png)
 
 The workflow of ``ocean_data_tools`` is to build uniform structs (e.g. ``argo``,
 ``cruise``, ``hycom``, ``mercator``, ``woa``, ``wod``) from raw datasets and 
@@ -47,14 +53,6 @@ is to continuously add support for additional datasets such as Remote Sensing
 Systems (RSS) and European Centre for Medium-Range 
 Weather Forecasts (ECMWF) products. The source code for ``ocean_data_tools`` has
 been archived to Zenodo with the linked DOI: [@Ferris:2020].
-
-![Building a virtual cruise from the Operational Mercator global ocean
-analysis and forecast system at 1/12 degree with 3D bathymetry [@Smith:1997]. 
-Showing (a) a 3D velocity plot created using ``model_domain_plot``, (b) 
-virtual cruise selection using ``transect_select``, and ``model_build_profiles``, 
-(c) coordinates of the resulting uniform struct, and (d) a temperature section 
-plotted using ``general_section`` with ``bathymetry_section``. Three of the 
-subplots use colormaps from cmocean [@Thyng:2016]. \label{fig:1}](figure.png)
 
 # Acknowledgements
 
