### 2020-07-03 b2dc8a2
Update paper.md
