### 2020-07-03 bf69232
Update paper.md
