+++ paper/paper.md
@@ -48,7 +48,7 @@ comparison of observations (such as those from underwater glider surveys) with
 model output. Some ``ocean_data_tools`` functions employ ``nctoolbox`` [@nctoolbox].
 
 ``ocean_data_tools`` has already been used in scientific publications [@Bemis:2020] 
-and [@Crear:2020]. The toolbox is built for extensibility, and the working plan 
+and [@Crear:2020]. This toolbox is built for extensibility, and the plan 
 is to continuously add support for additional datasets such as Remote Sensing 
 Systems (http://www.remss.com/) products and European Centre for Medium-Range 
 Weather Forecasts (ECMWF) products. The source code for ``ocean_data_tools`` has
