+++ paper/paper.bib
@@ -27,6 +27,13 @@
     doi = {10.5281/zenodo.3928714},
 }
 
+@article{McDougall:2011,
+  	Author = {{McDougall}, T. ~J. and {Barker}, P. ~M.},
+  	publisher = {SCOR/IAPSO WG127},
+  	Title = {{Getting started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox}~ISBN 978-0-646-55621-5},
+  	Year = 2011,
+}
+
 @misc{nctoolbox,
   	Author = {{Schlining}, B. and {Signell}, R. and {Crosby}, A.},
     publisher = {GitHub},
