+++ paper/paper.md
@@ -40,7 +40,8 @@ The workflow of ``ocean_data_tools`` is to build uniform structs (e.g. ``argo``,
 ``cruise``, ``hycom``, ``mercator``, ``woa``, ``wod``) from raw datasets and 
 call general functions on these structs to map, subset, or plot. Functions with 
 the ``\_build`` suffix load raw data into uniform structs. Structs are 
-compatible with all ``general_`` functions. One application of the ``\_build`` 
+compatible with all ``general_`` functions, as well as functions in the commonly-used
+Gibbs-SeaWater (GSW) Oceanographic Toolbox for MATLAB [@McDougall:2011]. One application of the ``\_build`` 
 feature is to create virtual cruises from model output \autoref{fig:1}. The user
 draws transects on a map (or passes coordinates as an argument) to build vertical profiles 
 from model data. This may be used as a cruise planning tool or to facilitate 
