+++ paper/paper.bib
@@ -27,11 +27,11 @@
     doi = {10.5281/zenodo.3928714},
 }
 
-@article{McDougall:2011,
+@book{McDougall:2011,
   	Author = {{McDougall}, T. ~J. and {Barker}, P. ~M.},
-  	publisher = {SCOR/IAPSO WG127},
-  	Title = {{Getting started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox}~ISBN 978-0-646-55621-5},
-  	Year = 2011,
+  	Title = {{Getting started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox ~ISBN 978-0-646-55621-5}},
+  	Publisher = {SCOR/IAPSO WG127},
+  	Year = 2011
 }
 
 @misc{nctoolbox,
