+++ paper/paper.bib
@@ -27,10 +27,10 @@
     doi = {10.5281/zenodo.3928714},
 }
 
-@book{McDougall:2011,
+@article{McDougall:2011,
   	Author = {{McDougall}, T. ~J. and {Barker}, P. ~M.},
   	Title = {{Getting started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox ~ISBN 978-0-646-55621-5}},
-  	Publisher = {SCOR/IAPSO WG127},
+  	Journal = {SCOR/IAPSO WG127},
   	Year = 2011
 }
 
