+++ paper/paper.bib
@@ -30,7 +30,7 @@
 @article{McDougall:2011,
   	Author = {{McDougall}, T. ~J. and {Barker}, P. ~M.},
   	Title = {{Getting started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox ~ISBN 978-0-646-55621-5}},
-  	Journal = {SCOR/IAPSO WG127},
+  	publisher = {SCOR/IAPSO WG127},
   	Year = 2011
 }
 
