+++ paper/paper.bib
@@ -29,8 +29,7 @@
 
 @article{McDougall:2011,
   	Author = {{McDougall}, T. ~J. and {Barker}, P. ~M.},
-  	Title = {{Getting started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox ~ISBN 978-0-646-55621-5}},
-  	publisher = {SCOR/IAPSO WG127},
+  	Title = {{Getting started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox, SCOR/IAPSO WG127, ISBN 978-0-646-55621-5}},
   	Year = 2011
 }
 
