+++ paper/paper.bib
@@ -20,7 +20,7 @@
 @article{Ferris:2020,
   	Author = {{Ferris}, L. ~N.},
   	Journal = {Zenodo},
-  	Title = {{lnferris/ocean_data_tools: ocean_data_tools_1.0.0}},
+  	Title = {{ocean_data_tools_1.0.0}},
   	Year = 2020,
     Month = Jul,
     Day = 02,
