### 2020-07-06 45bf06c
struct thinning
