### 2020-07-12 f466c30
better plotting
