### 2020-07-13 #2433 @kthyng
Hi @chadagreene and @castelao! Are you available to review this submission for JOSS? We're asking for reviews within 6 weeks, given the global pandemic. You can find more information about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
