+++ ocean_data_tools/utilities/model_simple_plot_velocity.m
@@ -1,11 +1,11 @@
 %  Author: Laur Ferris
 %  Email address: lnferris@alum.mit.edu
 %  Website: https://github.com/lnferris/ocean_data_tools
-%  Jun 2020; Last revision: 30-Jun-2020
+%  Jun 2020; Last revision: 13-Jul-2020
 %  Distributed under the terms of the MIT License
 %  Dependencies: nctoolbox
 
-function model_simple_plot_velocity(model,nc,date,region,depth,arrows)
+function [data,lat,lon] = model_simple_plot_velocity(model,nc,date,region,depth,arrows)
 
     if strcmp(model,'hycom')  
         sv = nc{'water_u'};     % Assign ncgeovariable handle.
@@ -60,17 +60,22 @@ if lonw > lone
 end
 
    if need2merge == 1
-        velocity = sqrt(double(sv.data(tin,din,lats:latn,lonw_A:lone_A)).^2+double(sv_v.data(tin,din,lats:latn,lonw_A:lone_A)).^2);
+        velocity_A = sqrt(double(sv.data(tin,din,lats:latn,lonw_A:lone_A)).^2+double(sv_v.data(tin,din,lats:latn,lonw_A:lone_A)).^2);
         figA = figure; % Plot left
-        pcolorjw(svg.lon(lonw_A:lone_A),svg.lat(lats:latn),velocity); % pcolorjw(x,y,c);    
+        pcolorjw(svg.lon(lonw_A:lone_A),svg.lat(lats:latn),velocity_A); % pcolorjw(x,y,c);    
 
-        velocity = sqrt(double(sv.data(tin,din,lats:latn,lonw_B:lone_B)).^2+double(sv_v.data(tin,din,lats:latn,lonw_B:lone_B)).^2);
+        velocity_B = sqrt(double(sv.data(tin,din,lats:latn,lonw_B:lone_B)).^2+double(sv_v.data(tin,din,lats:latn,lonw_B:lone_B)).^2);
         figB = figure; % Plot right
-        pcolorjw(svg.lon(lonw_B:lone_B)+360,svg.lat(lats:latn),velocity); % pcolorjw(x,y,c); 
+        pcolorjw(svg.lon(lonw_B:lone_B)+360,svg.lat(lats:latn),velocity_B); % pcolorjw(x,y,c); 
 
         merge_figures(figA,figB) % Merge left and right
         title({sprintf('velocity magnitude %.0fm',svg.z(din));datestr(svg.time(tin))},'interpreter','none');
         hcb = colorbar; title(hcb,sv.attribute('units'));
+        
+        % format optional output.
+        data = cat(2,squeeze(velocity_B),squeeze(velocity_A));
+        lat = svg.lat(lats:latn);
+        lon = [svg.lon(lonw_B:lone_B)+360; svg.lon(lonw_A:lone_A)];
 
         if  arrows==1
             u_left = squeeze(double(sv.data(tin,din,lats:latn,lonw_B:lone_B))); % Add directional arrows.
@@ -89,6 +94,11 @@ end
         pcolorjw(svg.lon(lonw:lone),svg.lat(lats:latn),velocity); % pcolorjw(x,y,c)
         title({sprintf('velocity magnitude %.0fm',svg.z(din));datestr(svg.time(tin))},'interpreter','none');
         hcb = colorbar; title(hcb,sv.attribute('units'));
+        
+        % format optional output
+        data = velocity;
+        lat = svg.lat(lats:latn);
+        lon = svg.lon(lonw:lone);
 
         if arrows==1
             u = squeeze(double(sv.data(tin,din,lats:latn,lonw:lone))); % Add directional arrows.
[…]
