+++ demos/demos.m
@@ -157,7 +157,7 @@ setup_nctoolbox
 
 bathymetry_dir = '/Users/lnferris/Documents/data/bathymetry/topo_20.1.nc'; % need to download
 
-% model_simple_plot - HYCOM EXAMPLE
+%% model_simple_plot - HYCOM EXAMPLE
 
 model = 'hycom'; % 'hycom' 'mercator'
 source = 'http://tds.hycom.org/thredds/dodsC/GLBv0.08/expt_57.7'; % url or local .nc 
@@ -166,7 +166,7 @@ variable = 'velocity';                  % 'water_u' 'water_v' 'water_temp' 'sali
 region = [-5.0, 45.0 ,160,-150 ];      % [-90 90 -180 180]
 depth = -150;                          % Depth level between 0 and -5000m
 arrows = 0;                            % Velocity direction arrows 1=on 0=off
-model_simple_plot(model,source,date,variable,region,depth,arrows)
+[data,lat,lon] = model_simple_plot(model,source,date,variable,region,depth,arrows); % optionally output the plotted data layer
 
 % model_build_profiles  - HYCOM EXAMPLE
 

[…]
