+++ ocean_data_tools/woa_domain_plot.m
@@ -61,26 +61,25 @@ else
 end
 
 % Make Plot
+
+figure 
+    
 if region(3) > region(4) && region(4) < 0 % If data spans the dateline...
     
-    figure; % Plot left side
-    data = reshape(permute(squeeze(double(sv.data(1,:,lats:latn,lonw_A:lone_A))),[2,3,1]),[],1); % permute array to be lon x lat x dep
+    dataA = reshape(permute(squeeze(double(sv.data(1,:,lats:latn,lonw_A:lone_A))),[2,3,1]),[],1); % permute array to be lon x lat x dep
     [lon_mesh,lat_mesh,dep_mesh] = meshgrid(svg.lon(lonw_A:lone_A),svg.lat(lats:latn),svg.z(:)); 
     lon = reshape(lon_mesh,[],1); 
     lat = reshape(lat_mesh,[],1); 
     z = reshape(dep_mesh,[],1);
-    scatter3(lon,lat,z,[],data,'.')
+    scatter3(lon,lat,z,[],dataA,'.')
     
-    hold on % Plot right side
-    data = reshape(permute(squeeze(double(sv.data(1,:,lats:latn,lonw_B:lone_B))),[2,3,1]),[],1); % permute array to be lon x lat x dep
+    hold on 
+    dataB = reshape(permute(squeeze(double(sv.data(1,:,lats:latn,lonw_B:lone_B))),[2,3,1]),[],1); % permute array to be lon x lat x dep
     [lon_mesh,lat_mesh,dep_mesh] = meshgrid(svg.lon(lonw_B:lone_B)+360,svg.lat(lats:latn),svg.z(:)); 
     lon = reshape(lon_mesh,[],1); 
     lat = reshape(lat_mesh,[],1); 
     z = reshape(dep_mesh,[],1);
-    scatter3(lon,lat,z,[],data,'.')
-
-    title({sprintf('%s',sv.attribute('standard_name'))},'interpreter','none');
-    hcb = colorbar; title(hcb,sv.attribute('units'));
+    scatter3(lon,lat,z,[],dataB,'.')
       
 else   
 
@@ -89,12 +88,11 @@ else
     lon = reshape(lon_mesh,[],1); 
     lat = reshape(lat_mesh,[],1); 
     z = reshape(dep_mesh,[],1);
-
-    figure 
     scatter3(lon,lat,z,[],data,'.')
-    title({sprintf('%s',sv.attribute('standard_name'))},'interpreter','none');
-    hcb = colorbar; title(hcb,sv.attribute('units'));
     
 end  
+
+title({sprintf('%s',sv.attribute('standard_name'))},'interpreter','none');
+hcb = colorbar; title(hcb,sv.attribute('units'),'interpreter','none');
 
 end
