+++ ocean_data_tools/utilities/model_domain_plot_velocity.m
@@ -62,30 +62,28 @@ end
 
 % Make Plot
 
+figure 
+
 if need2merge == 1
     
-    figure; % Plot left side
     u = reshape(permute(squeeze(double(sv.data(tin,:,lats:latn,lonw_A:lone_A))),[2,3,1]),[],1); % permute array to be lon x lat x dep
     v = reshape(permute(squeeze(double(sv_v.data(tin,:,lats:latn,lonw_A:lone_A))),[2,3,1]),[],1); % permute array to be lon x lat x dep
-    data = sqrt(u.^2+v.^2);
+    dataA = sqrt(u.^2+v.^2);
     [lon_mesh,lat_mesh,dep_mesh] = meshgrid(svg.lon(lonw_A:lone_A),svg.lat(lats:latn),svg.z(:)); 
     lon = reshape(lon_mesh,[],1); 
     lat = reshape(lat_mesh,[],1); 
     z = reshape(dep_mesh,[],1);
-    scatter3(lon,lat,z,[],data,'.')
+    scatter3(lon,lat,z,[],dataA,'.')
     
-    hold on % Plot right side
+    hold on 
     u = reshape(permute(squeeze(double(sv.data(tin,:,lats:latn,lonw_B:lone_B))),[2,3,1]),[],1); % permute array to be lon x lat x dep
     v = reshape(permute(squeeze(double(sv_v.data(tin,:,lats:latn,lonw_B:lone_B))),[2,3,1]),[],1); % permute array to be lon x lat x dep
-    data = sqrt(u.^2+v.^2);
+    dataB = sqrt(u.^2+v.^2);
     [lon_mesh,lat_mesh,dep_mesh] = meshgrid(svg.lon(lonw_B:lone_B)+360,svg.lat(lats:latn),svg.z(:)); 
     lon = reshape(lon_mesh,[],1); 
     lat = reshape(lat_mesh,[],1); 
     z = reshape(dep_mesh,[],1);
-    scatter3(lon,lat,z,[],data,'.')
-    
-    title({'velocity magnitude';datestr(svg.time(tin))},'interpreter','none');
-    hcb = colorbar; title(hcb,sv.attribute('units'));
+    scatter3(lon,lat,z,[],dataB,'.')
 
 else   
     u = reshape(permute(squeeze(double(sv.data(tin,:,lats:latn,lonw:lone))),[2,3,1]),[],1); % permute array to be lon x lat x dep
@@ -95,12 +93,11 @@ else
     lon = reshape(lon_mesh,[],1); 
     lat = reshape(lat_mesh,[],1); 
     z = reshape(dep_mesh,[],1);
-
-    figure 
     scatter3(lon,lat,z,[],data,'.')
-    title({'velocity magnitude';datestr(svg.time(tin))},'interpreter','none');
-    hcb = colorbar; title(hcb,sv.attribute('units'));
+
 end    
 
+title({'velocity magnitude';datestr(svg.time(tin))},'interpreter','none');
+hcb = colorbar; title(hcb,sv.attribute('units'),'interpreter','none');
 
 end
