### 2020-07-14 #2433 @castelao
Hi @kthyng , I'm not a regular Matlab user for a long time now but I still do a few bits when necessary. Therefore, I won't be able to evaluate it as a modern Matlab way of doing things. If that's fine, I can review this submission.
