### 2020-07-20 #2433 @kthyng
@lnferris I know you've asked questions about your submission in a pre-submission inquiry, so here is a little more info. Our review process is usually fairly interactive between the author(s) and reviewers, and it can be a great opportunity to learn code/package best practices, so I encourage you to take advantage of your reviewer expertise! Your package looks useful, though a little thin, so if there are clear places that come up in review to add in some additional functionality, that could make your package used quite a bit more. Let me know if you have questions along the way!
