### 2020-07-20 #2433 @kthyng
@chadagreene This is excellent to have as a starting point in the meantime! Could you start an issue in the [repo](https://github.com/lnferris/ocean_data_tools) with this comment, and provide a link into the [review issue](https://github.com/openjournals/joss-reviews/issues/2497)? This is the best way to keep track of issues that come up in the review so that details are linked in but not taking up the review issue space.
