### 2020-07-20 #2433 @chadagreene
@lnferris Sounds good.

Sorry to bread-crumb the suggestions here, but you might also want to provide some sort overview page, where all the functions are organized in a way that users can quickly find the function they need based on the task they're trying to accomplish. I'm thinking [something along these lines](https://www.chadagreene.com/CDT/CDT_Contents.html), and it could take the form of a documentation page accessible in the Matlab Documentation Viewer, or it could just be detailed in the README.md. Or perhaps you'll come up with some other way to do it, but you know, just something to help users figure out which functions they need for a given task.
