+++ ocean_data_tools/argo_build.m
@@ -1,11 +1,36 @@
 
-%  Author: Laur Ferris
-%  Email address: lnferris@alum.mit.edu
-%  Website: https://github.com/lnferris/ocean_data_tools
-%  Jun 2020; Last revision: 27-Jun-2020
-%  Distributed under the terms of the MIT License
-
 function [argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)
+% argo_build searches path argo_dir for relevant profiles and loads them
+% into struct argo.
+% 
+%% Syntax
+% 
+%  [argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)
+% 
+%% Description 
+% 
+% argo_build searches pathway argo_dir for profiles meeting the search criteria
+% region, start_date, and end_date. Profiles are loaded into the struct
+% array argo with all variables specified in variable_list. Variables PLATFORM_NUMBER, 
+% LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically.
+% Files containing matching profiles are listed in matching_files.
+% 
+%% Example 1
+% Load Argo data from west of New Zealand 
+% 
+% argo_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/argo/*profiles*.nc';
+% region = [-60.0 -50.0 150.0 160.0]; %  Search region [-90 90 -180 180]
+% start_date = '01-Nov-2015 00:00:00';
+% end_date = '01-Jan-2017 00:00:00';
+% variable_list = {'TEMP_ADJUSTED','PSAL_ADJUSTED'};
+% [argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list);
+%
+%% Citation Info 
+% github.com/lnferris/ocean_data_tools
+% Jun 2020; Last revision: 27-Jun-2020
+% 
+% See also argo_platform_subset and general_region_subset.
+
 
     FillValue = 99999; % From Argo manual.
     start_date = datenum(start_date,'dd-mmm-yyyy HH:MM:SS');
