### 2020-07-20 #2433 @kthyng
@chadagreene Again, these suggestions are great, but please put them each in a separate issue on the github repo, then provide  a link in the issue to the review issue (which is not this page). Nothing else should go here in the pre-review issue. Let me know if you need any help with this part of the JOSS review since it's unusual.
