+++ ocean_data_tools/argo_platform_map.m
@@ -1,10 +1,32 @@
-%  Author: Laur Ferris
-%  Email address: lnferris@alum.mit.edu
-%  Website: https://github.com/lnferris/ocean_data_tools
-%  Jun 2020; Last revision: 16-Jun-2020
-%  Distributed under the terms of the MIT License
 
 function argo_platform_map(argo,annotate)
+% argo_platform_map plots locations of Argo profiles in struct argo, coloring markers by platform
+% 
+%% Syntax
+% 
+%  argo_platform_map(argo)
+%  argo_platform_map(argo,annotate)
+% 
+%% Description 
+% 
+% argo_platform_map plots locations of Argo profiles in argo,
+% coloring markers by Argo platform; where argo is a struct created by argo_build.
+%  
+% argo_platform_map(argo,annotate) adds number annotations to the markers. by default annotate=0.
+% set annotate=1 to turn on annotation.
+% 
+%% Example 1
+% Plot locations of the profiles in struct argo
+% 
+% annotate = 1; % 1=on 0=off
+% argo_platform_map(argo,annotate)
+%
+%
+%% Citation Info 
+% github.com/lnferris/ocean_data_tools
+% 
+% See also argo_build and argo_platform_subset.
+
 
     more_colors()
 
@@ -21,7 +43,7 @@ function argo_platform_map(argo,annotate)
         plot(argo.lon(argo.id==platformids(i)),argo.lat(argo.id==platformids(i)),'.','MarkerSize',14)
 
         if nargin == 2 && annotate ==1
-            text(argo.lon(argo.id==platformids(i)),argo.lat(argo.id==platformids(i)),string(x(i)),'FontSize',6)
+            text(argo.lon(argo.id==platformids(i)),argo.lat(argo.id==platformids(i)),string(x(i)),'FontSize',10)
         end
 
     end
