+++ ocean_data_tools/argo_platform_map.m
@@ -9,14 +9,14 @@ function argo_platform_map(argo,annotate)
 % 
 %% Description 
 % 
-% argo_platform_map plots locations of Argo profiles in argo,
+% argo_platform_map(argo) plots locations of Argo profiles in argo,
 % coloring markers by Argo platform; where argo is a struct created by argo_build.
 %  
 % argo_platform_map(argo,annotate) adds number annotations to the markers. by default annotate=0.
 % set annotate=1 to turn on annotation.
 % 
 %% Example 1
-% Plot locations of the profiles in struct argo
+% Plot locations of the profiles in struct argo:
 % 
 % annotate = 1; % 1=on 0=off
 % argo_platform_map(argo,annotate)
@@ -24,6 +24,7 @@ function argo_platform_map(argo,annotate)
 %
 %% Citation Info 
 % github.com/lnferris/ocean_data_tools
+% Jun 2020; Last revision: 20-Jul-2020
 % 
 % See also argo_build and argo_platform_subset.
 
