+++ ocean_data_tools/argo_profiles_map.m
@@ -1,10 +1,35 @@
-%  Author: Laur Ferris
-%  Email address: lnferris@alum.mit.edu
-%  Website: https://github.com/lnferris/ocean_data_tools
-%  Jun 2020; Last revision: 16-Jun-2020
-%  Distributed under the terms of the MIT License
 
 function argo_profiles_map(argo,annotate)
+% argo_profiles_map plots locations of Argo profiles in struct argo,
+% coloring markers by profile.
+% 
+%% Syntax
+% 
+%  argo_profiles_map(argo)
+%  argo_profiles_map(argo,annotate)
+% 
+%% Description 
+%  
+% argo_profiles_map(argo) plots locations of Argo profiles in struct argo,
+% coloring markers by profile; where argo is a struct created by argo_build.
+% The colors of profiles corresponds to those of argo_profiles called on the
+% same struct.
+
+% argo_profiles_map(argo,annotate) adds number annotations to the markers. by default annotate=0.
+% set annotate=1 to turn on annotation. The annotations of profiles correspond to
+% those of argo_profiles called on the same struct.
+% 
+%% Example 1
+% Plot locations of the profiles in struct argo:
+% 
+% annotate = 1; 
+% argo_profiles_map(argo,annotate) % annotate optional,  1=on 0=off
+%
+%% Citation Info 
+% github.com/lnferris/ocean_data_tools
+% Jun 2020; Last revision: 16-Jun-2020
+% 
+% See also argo_build and argo_profiles.
 
     more_colors()
 
