### 2020-07-21 #2497 @chadagreene
Hmm, I officially accepted the invitation to review a couple hours ago, and now my email is getting flooded by emails with the subject line starting

Re: [openjournals/joss-reviews] [PRE REVIEW]:

and they are all unrelated to the review I accepted the invitation for. I have hit "unsubscribe" on about 10 of these emails so far, yet they keep rolling in. Anybody know how I can stop receiving these emails?
