+++ demos/demos.m
@@ -231,16 +231,12 @@ variable_list = {'salinity','temperature','oxygen'};
 [cruise] = whp_cruise_build(ctdo_dir,uv_dir,wvke_dir,variable_list); % Use a dummy path (e.g. uv_dir ='null') if missing data. 
 general_map(cruise,bathymetry_dir,'2Dcontour')
 
-% whp_cruise_section
-
 variable = 'temperature'; % See cruise for options.
 xref = 'lon'; % See cruise for options.
 zref = 'pressure'; % See cruise for options.
 interpolate = 1; % 1=on 0=off
 contours = 0; % 1=on 0=off
 general_section(cruise,variable,xref,zref,interpolate,contours) % interpolate, contours optional
-
-% whp_cruise_profiles
 
 general_profiles(cruise,variable,zref)
 
