+++ README.md
@@ -17,7 +17,7 @@ nctoolbox (https://github.com/nctoolbox/nctoolbox)
 ## Getting Started
 
 1. Download bathymetry (see below).
-2. Download the nctoolbox (https://github.com/nctoolbox/nctoolbox).
+2. Download the nctoolbox (https://github.com/nctoolbox/nctoolbox). Remember to run the command ``setup_nctoolbox`` before use.
 3. Run each demonstration in **demos/demos.m**.
 
 Functions are named using a two-part system. The prefix (``argo_``, ``bathymetry_``, ``general_``, etc.) indicates the appropriate data source, while the suffix (``\_build``, ``\_profiles``, ``\_section``, etc.) indicates the action performed. Functions with the ``\_build`` suffix load raw data into uniform structs (e.g. ``argo``, ``cruise``, ``hycom``, ``mercator``, ``woa``, ``wod``). Uniform structs created by ``\_build`` functions are compatable with any ``general_`` function, as well as functions in the Gibbs-SeaWater (GSW) Oceanographic Toolbox (http://www.teos-10.org/software.htm#1).
