+++ ocean_data_tools/utilities/merge_figures.m
@@ -1,4 +1,5 @@
 function merge_figures(figA,figB)
+% merge two figures
     copyobj(get(get(figB, 'Children'),'Children'), get(figA, 'Children')); % Merge the two figures.
     close(figB)
     axis tight;
