+++ ocean_data_tools/utilities/more_colors.m
@@ -1,6 +1,6 @@
-% expand default color order
 
 function more_colors()
+% expand default color order
 c_order = [0 .447 .741;.85 .325 .098; .929 .694 .125;0 1 0;.494 .184 .556;...
       .466 .674 .188;1 .7 .7; .301 .745 .933;0 0 0; .635 .078 .184;0 .5 0;...
       0 .75 .75;.75 0 .75;.75 .75 0;1 .5 .8;.25 .25 .25;1 0 1;0 1 1;1 0 0;...
