### 2020-07-22 #2433 @lnferris
I added ``help`` accessible added headers (Syntax, Description, Examples, Info) to all functions. 👍
