+++ README.md
@@ -58,3 +58,6 @@ Functions build the World Ocean Atlas url based on arguments, but coarser resolu
 ### wod_
 
 Search and select World Ocean Database data at https://www.nodc.noaa.gov/OC5/SELECT/dbsearch/dbsearch.html.
+
+## Contents
+
