+++ README.md
@@ -129,7 +129,7 @@ Search and select World Ocean Database data at https://www.nodc.noaa.gov/OC5/SEL
 
 **netcdf_info** gets information about the first netcdf file in a path and saves it to a text file
 
-**region_select** creates coordinate list (which represents vertices of a polygon) by clicking stations on a plot
+**region_select** creates coordinate list (which represents vertices of a polygon region) by clicking stations on a plot
 
-**transect_select** creates a coordinate list (which represents a transect) by clicking stations on a plot
+**transect_select** creates a coordinate list (which represents a virtual transect) by clicking stations on a plot
 
