+++ documentation/example00.md
@@ -1,0 +1,2 @@
+# example00
+
