+++ docs/argo_build.md
@@ -1,4 +1,4 @@
-#### argo_build
+### argo_build
 
 
 #### Syntax
