+++ docs/argo_build.md
@@ -1,4 +1,4 @@
-#### argo_build
+### argo_build
 
 
 #### Syntax
@@ -37,4 +37,6 @@ general_map(argo,bathymetry_dir,'2Dcontour')
 ```
 <img src="https://user-images.githubusercontent.com/24570061/88301724-fd1dab80-ccd2-11ea-9ea7-7badf1424865.png" width="700">
 <img src="https://user-images.githubusercontent.com/24570061/88301788-11fa3f00-ccd3-11ea-9cdf-1622f701bfe9.png" width="700">
+
+[Back](https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1)
 
