+++ docs/argo_build.md
@@ -1,5 +1,6 @@
 ### argo_build
 
+[Back to README](https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1)
 
 #### Syntax
 
@@ -37,6 +38,3 @@ general_map(argo,bathymetry_dir,'2Dcontour')
 ```
 <img src="https://user-images.githubusercontent.com/24570061/88301724-fd1dab80-ccd2-11ea-9ea7-7badf1424865.png" width="700">
 <img src="https://user-images.githubusercontent.com/24570061/88301788-11fa3f00-ccd3-11ea-9cdf-1622f701bfe9.png" width="700">
-
-[Back to README](https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1)
-
