+++ docs/argo_build.md
@@ -1,6 +1,6 @@
-### argo_build
+[Go back](https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1)
 
-[Back to README](https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1)
+### argo_build
 
 #### Syntax
 
