### 2020-07-23 #2497 @kthyng
@chadagreene please see the issues I started on your behalf above, and create similar issues for future comments. Then usually there is a short summary of a review actually in this review issue that refers to the open issues. There is a bit of bookkeeping with this setup.

@lnferris Can you copy and paste your comments that have been in response to these comments from @chadagreene into the relevant issue above? This way, it is clear what work has been done on a given issue, as well as it being clear when it is finished since the issue can be closed.
