### 2020-07-23 lnferris/ocean_data_tools#2 @lnferris
I added ``help``-accessible added headers (Syntax, Description, Examples, Info) to all functions. 👍
