### 2020-07-23 lnferris/ocean_data_tools#3 @lnferris
@chadagreene I have added Contents to README and am implementing documentation using relative links between markdown files e.g. https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1 links to https://github.com/lnferris/ocean_data_tools/blob/master/docs/argo_build.md
