### 2020-07-23 5c79aa9
Update argo_platform_map.md
