### 2020-07-23 e74ea06
Update argo_platform_map.md
