### 2020-07-23 d20e13e
Update argo_platform_map.md
