+++ docs/argo_platform_map.md
@@ -39,5 +39,5 @@ bathymetry_plot(bathymetry_dir,bathymetry_region(argo),'2Dcontour')
 ```
 <img src="https://user-images.githubusercontent.com/24570061/88316847-6955da80-cce6-11ea-8bb0-d9d0523a3a29.png" width="700">
 
-[Back](https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1)
+[Back](https://github.com/lnferris/ocean_data_tools#additional-functions-for-inspecting-argo-data-1)
 
