### 2020-07-23 bad1a12
Update argo_platform_subset.md
