+++ docs/argo_platform_subset.md
@@ -3,11 +3,11 @@
 #### Syntax
 
 ```Matlab
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)
+[subargo] = argo_platform_subset(argo,platform_id) 
 ```
 #### Description
 
-``[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)`` searches pathway ``argo_dir`` for profiles meeting the search criteria ``region``, ``start_date``, and ``end_date``. Profiles are loaded into the struct array ``argo`` with all variables specified in ``variable_list``. Variables PLATFORM_NUMBER, LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically. Files containing matching profiles are listed in ``matching_files``.
+``[subargo] = argo_platform_subset(argo,platform_id)`` subsets ``argo`` by Argo platform ID (PLATFORM_NUMBER) into struct ``subargo``; where ``argo`` is a struct created by ``argo_build`` and platform_id is the integer ID
 
 
 #### Example 1
@@ -28,14 +28,23 @@ end_date = '01-Jan-2017 00:00:00';
 variable_list = {'TEMP_ADJUSTED','PSAL_ADJUSTED'};
 [argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list);
 
-% Make plots:
+% Subset by platform:
 
-general_profiles(argo,variable,'depth')
-general_map(argo,bathymetry_dir,'2Dcontour')
+platform_id = 5904421;
+[subargo] = argo_platform_subset(argo,platform_id);
+
+% Make a plot:
+
+argo_platform_map(subargo,1)
+bathymetry_plot(bathymetry_dir,bathymetry_region(argo),'2Dcontour')
 
 ```
-<img src="https://user-images.githubusercontent.com/24570061/88301724-fd1dab80-ccd2-11ea-9ea7-7badf1424865.png" width="600">
-<img src="https://user-images.githubusercontent.com/24570061/88301788-11fa3f00-ccd3-11ea-9cdf-1622f701bfe9.png" width="600">
+<img src="https://user-images.githubusercontent.com/24570061/88316847-6955da80-cce6-11ea-8bb0-d9d0523a3a29.png" width="700">
+
+
+<img src="https://user-images.githubusercontent.com/24570061/88324607-ec306280-ccf1-11ea-8f9a-81320046ccf4.png" width="700">
+
+
 
 [Back](https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1)
 
