### 2020-07-23 4fee4c1
Update argo_platform_subset.md
