### 2020-07-23 f091054
Update argo_platform_subset.md
