### 2020-07-23 8347602
Update argo_platform_subset.md
