### 2020-07-23 4bdfba5
Update argo_platform_subset.md
