### 2020-07-23 bcbfdf6
Update mocha_simple_plot.md
