### 2020-07-23 37e1fae
Update mocha_simple_plot.md
