+++ docs/mocha_simple_plot.md
@@ -7,7 +7,7 @@ mocha_simple_plot(month,depth,variable,region)
 ```
 #### Description
 
-``mocha_simple_plot(month,depth,variable,region)`` plots the nearest available depth-level to ``depth``. ``variable`` specifies the parameter to be plotted and ``region`` is the rectangular region to be plotted.
+``mocha_simple_plot(month,depth,variable,region)`` plots the nearest available depth-level to ``depth``. ``variable`` specifies the parameter to be plotted and ``region`` is the rectangular region to be plotted. The calendar month is specified by ``month``.
 
 #### Example 1
 
