### 2020-07-23 e26edac
Update mocha_domain_plot.md
