+++ docs/mocha_domain_plot.md
@@ -1,41 +1,39 @@
-### argo_build
+### mocha_build_profiles
 
 #### Syntax
 
 ```Matlab
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)
+mocha_domain_plot(month,variable,region)
 ```
 #### Description
 
-``[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)`` searches pathway ``argo_dir`` for profiles meeting the search criteria ``region``, ``start_date``, and ``end_date``. Profiles are loaded into the struct array ``argo`` with all variables specified in ``variable_list``. Variables PLATFORM_NUMBER, LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically. Files containing matching profiles are listed in ``matching_files``.
-
+``mocha_domain_plot(month,variable,region)`` plots all depth-levels of ``variable`` over the specified ``region``. The calendar month is specified by ``month``.
 
 #### Example 1
 
 
 ```Matlab
 
-% Get variable information:
+% Setup nctoolbox:
 
-argo_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/argo/*profiles*.nc';
-netcdf_info(argo_dir);
+setup_nctoolbox
 
-% Load Argo data from west of New Zealand:
+% Plot a 3-D temperature domain:
 
-region = [-60.0 -50.0 150.0 160.0]; %  Search region [-90 90 -180 180]
-start_date = '01-Nov-2015 00:00:00';
-end_date = '01-Jan-2017 00:00:00';
-variable_list = {'TEMP_ADJUSTED','PSAL_ADJUSTED'};
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list);
+month = 10; % Month (1 through 12).
+variable = 'temperature'; %  'temperature' 'salinity'
+region = [34 42  -80 -70]; % [30 48 -80 -58]
+mocha_domain_plot(month,variable,region)
 
-% Make plots:
-
-general_profiles(argo,variable,'depth')
-general_map(argo,bathymetry_dir,'2Dcontour')
+% Add bathymetry:
+bathymetry_dir = '/Users/lnferris/Documents/data/bathymetry/topo_20.1.nc';
+bathymetry_plot(bathymetry_dir,region,'3Dsurf')
+caxis([27 38])
 
 ```
-<img src="https://user-images.githubusercontent.com/24570061/88301724-fd1dab80-ccd2-11ea-9ea7-7badf1424865.png" width="600">
-<img src="https://user-images.githubusercontent.com/24570061/88301788-11fa3f00-ccd3-11ea-9cdf-1622f701bfe9.png" width="600">
 
-[Back](https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1)
+<img src="https://user-images.githubusercontent.com/24570061/88339261-ba29fb00-cd07-11ea-97a3-bef5868a1fa4.png" width="600">
+
+
+[Back](https://github.com/lnferris/ocean_data_tools#plotting-gridded-data-without-building-structs-1)
 
