### 2020-07-23 48517a5
Update mocha_domain_plot.md
