### 2020-07-23 27bbd77
Update whp_cruise_build.md
