### 2020-07-23 eb23646
Update whp_cruise_build.md
