### 2020-07-23 775c727
Update whp_cruise_build.md
