### 2020-07-23 f55b4d1
Update whp_cruise_build.md
