### 2020-07-23 d6db1ac
Update whp_cruise_build.md
