### 2020-07-23 8d9b86a
Update whp_cruise_build.md
