### 2020-07-23 56cda5c
Update whp_cruise_build.md
