### 2020-07-24 94fe1cc
Update woa_build_profiles.md
