+++ docs/woa_build_profiles.md
@@ -1,13 +1,27 @@
-### argo_build
+### woa_build_profiles
 
 #### Syntax
 
 ```Matlab
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)
+[woa] =   woa_build_profiles(variable_list,time,xcoords,ycoords)
+[woa] =   woa_build_profiles(variable_list,time,xcoords,ycoords,zgrid)
 ```
 #### Description
 
-``[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)`` searches pathway ``argo_dir`` for profiles meeting the search criteria ``region``, ``start_date``, and ``end_date``. Profiles are loaded into the struct array ``argo`` with all variables specified in ``variable_list``. Variables PLATFORM_NUMBER, LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically. Files containing matching profiles are listed in ``matching_files``.
+``[woa] =   woa_build_profiles(variable_list,time,xcoords,ycoords)`` builds a struct of profiles from World Ocean Atlas 2018 Statistical Mean for All Decades Objectively Analyzed Mean Fields at Standard Depth Levels, pulling profiles nearest to coordinates specified by ``xcoords`` and ``ycoords``. ``time`` specifies monthly or annual climatology; ``time ='00'`` for annual climatology and ``'01'`` ``'10'`` etc. for monthly climatology. Profiles are loaded into the struct array ``woa`` with all variables specified in ``variable_list``. The function builds the url, extracting the maximum resolution available (typically 0.25-deg or 1.00-degree grid). Resolution depends on the variable. If the user requests only 0.25-degree variables in variable_list, data will be returned in  0.25-degree resolution. If any requested variable is coarser (1-degree) all variables will be returned in 1-degree resolution.
+
+Variables are:
+
+``'temperature'`` (degrees Celsius)    
+``'salinity'`` (psu)                    
+``'oxygen'`` (umol/kg)                 
+``'o2sat'`` (%)
+``'AOU'`` (umol/kg)                
+``'silicate'`` (umol/kg)          
+``'phosphate'`` (umol/kg)   
+``'nitrate'`` (umol/kg)                      
+
+``[woa] =  woa_build_profiles(variable_list,time,xcoords,ycoords,zgrid)`` depth-interpolates the profiles to a vertical grid of ``zgrid``, in meters. ``zgrid=2`` would produce profiles interpolated to 2 meter vertical grid.
 
 
 #### Example 1
@@ -15,27 +29,41 @@
 
 ```Matlab
 
-% Get variable information:
+% Setup nctoolbox:
 
-argo_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/argo/*profiles*.nc';
-netcdf_info(argo_dir);
+setup_nctoolbox
 
-% Load Argo data from west of New Zealand:
+% Plot surface nitrate:
 
-region = [-60.0 -50.0 150.0 160.0]; %  Search region [-90 90 -180 180]
-start_date = '01-Nov-2015 00:00:00';
-end_date = '01-Jan-2017 00:00:00';
-variable_list = {'TEMP_ADJUSTED','PSAL_ADJUSTED'};
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list);
+variable = 'nitrate';
+time = '03';
+region = [-5.0, 45.0 ,-120, -150]; 
+depth = -0; 
+woa_simple_plot(variable,time,region,depth)
 
-% Make plots:
+% Click stations on the plot to create a coordinate list:
 
[…]
