### 2020-07-24 1224ad2
Update woa_build_profiles.md
