+++ docs/woa_build_profiles.md
@@ -10,6 +10,8 @@
 
 ``[woa] =   woa_build_profiles(variable_list,time,xcoords,ycoords)`` builds a struct of profiles from World Ocean Atlas 2018 Statistical Mean for All Decades Objectively Analyzed Mean Fields at Standard Depth Levels, pulling profiles nearest to coordinates specified by ``xcoords`` and ``ycoords``. ``time`` specifies monthly or annual climatology; ``time ='00'`` for annual climatology and ``'01'`` ``'10'`` etc. for monthly climatology. Profiles are loaded into the struct array ``woa`` with all variables specified in ``variable_list``. The function builds the url, extracting the maximum resolution available (typically 0.25-deg or 1.00-degree grid). Resolution depends on the variable. If the user requests only 0.25-degree variables in variable_list, data will be returned in  0.25-degree resolution. If any requested variable is coarser (1-degree) all variables will be returned in 1-degree resolution.
 
+``[woa] =  woa_build_profiles(variable_list,time,xcoords,ycoords,zgrid)`` depth-interpolates the profiles to a vertical grid of ``zgrid``, in meters. ``zgrid=2`` would produce profiles interpolated to 2 meter vertical grid.
+
 Available variables are:
 
 ``'temperature'`` (degrees Celsius)    
@@ -21,9 +23,6 @@ Available variables are:
 ``'silicate'`` (umol/kg)          
 ``'phosphate'`` (umol/kg)   
 ``'nitrate'`` (umol/kg)                      
-
-``[woa] =  woa_build_profiles(variable_list,time,xcoords,ycoords,zgrid)`` depth-interpolates the profiles to a vertical grid of ``zgrid``, in meters. ``zgrid=2`` would produce profiles interpolated to 2 meter vertical grid.
-
 
 #### Example 1
 
