### 2020-07-24 abad2b2
Update woa_build_profiles.md
