### 2020-07-24 4d50921
Update wod_build.md
