+++ README.md
@@ -170,6 +170,6 @@ Search the [World Ocean Database](https://www.nodc.noaa.gov/OC5/SELECT/dbsearch/
 
 **transect_select** creates a coordinate list (which represents a virtual transect) by clicking stations on a plot
 
-<img width="777" alt="Screen Shot 2020-07-22 at 23 50 25" src="https://user-images.githubusercontent.com/24570061/88250639-2b20d280-cc76-11ea-9c94-3ce16300f735.png">
+<img width="1000" alt="Screen Shot 2020-07-22 at 23 50 25" src="https://user-images.githubusercontent.com/24570061/88250639-2b20d280-cc76-11ea-9c94-3ce16300f735.png">
 
 
