+++ README.md
@@ -10,17 +10,13 @@ ocean_data_tools simplifies the process of extracting, formatting, and visualizi
 
 Find a bug, have a question, or want to contribute? Open an issue or email lnferris@alum.mit.edu.
 
-### [Dependencies](#dependencies-1)
-
 ### [Getting Started](#getting-started-1)
+
+### [Dependencies](#dependencies-1)
 
 ### [Contents](#contents-1)
 
 ### [Finding Data](#finding-data-1)
-
-## Dependencies
-
-[nctoolbox](https://github.com/nctoolbox/nctoolbox)
 
 ## Getting Started
 
@@ -31,6 +27,10 @@ Find a bug, have a question, or want to contribute? Open an issue or email lnfer
 Functions are named using a two-part system. The prefix (``argo_``, ``bathymetry_``, ``general_``, etc.) indicates the appropriate data source, while the suffix (``\_build``, ``\_profiles``, ``\_section``, etc.) indicates the action performed. Functions with the ``\_build`` suffix load raw data into uniform structs (e.g. ``argo``, ``cruise``, ``hycom``, ``mercator``, ``woa``, ``wod``). Uniform structs created by ``\_build`` functions are compatable with any ``general_`` function, as well as functions in the [Gibbs-SeaWater (GSW) Oceanographic Toolbox](http://www.teos-10.org/software.htm#1).
 
 Main functions are located in **ocean_data_tools/**. Demonstrations are located in **demos/**. Test datas are located in **data/**. Shell scripts for batch downloading data are located in **shell_scripts/**. While shell scripts can be run directly in a macOS Terminal, running them in Windows requires [Cygwin](https://www.cygwin.com/) or similar. Python syntax examples are located in **python/**, which may be grow to become a module in the future.
+
+## Dependencies
+
+[nctoolbox](https://github.com/nctoolbox/nctoolbox)
 
 ## Contents
 
