### 2020-07-24 fad5ebf
Update gsw_example.md
