### 2020-07-24 0523ee1
Update gsw_example.md
