### 2020-07-24 f74bd39
Update gsw_example.md
