### 2020-07-24 af9483d
Update gsw_example.md
