### 2020-07-24 3275fce
Update gsw_example.md
