### 2020-07-24 9054e96
Update transect_select.md
