+++ docs/transect_select.md
@@ -1,13 +1,16 @@
-### argo_build
+### transect_select
 
 #### Syntax
 
 ```Matlab
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)
+[xcoords,ycoords] = transect_select() 
+[xcoords,ycoords] = transect_select(densify)
 ```
 #### Description
 
-``[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)`` searches pathway ``argo_dir`` for profiles meeting the search criteria ``region``, ``start_date``, and ``end_date``. Profiles are loaded into the struct array ``argo`` with all variables specified in ``variable_list``. Variables PLATFORM_NUMBER, LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically. Files containing matching profiles are listed in ``matching_files``.
+``[xcoords,ycoords] = transect_select()`` creates a list of x and y coordinates selected by clicking stations on an existing (latitude vs. longitude) plot, returning them as ``xcoords`` and ``ycoords``
+
+``[xcoords,ycoords] = transect_select(densify)`` auto-generates additional stations with the multiplier ``densify``. ``densify=10`` would fill in 10 stations for every station clicked using linear interpolation of complex coordinates.
 
 
 #### Example 1
@@ -15,27 +18,38 @@
 
 ```Matlab
 
-% Get variable information:
+% Plot HYCOM surface salinity:
 
-argo_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/argo/*profiles*.nc';
-netcdf_info(argo_dir);
+model = 'hycom'; 
+source = 'http://tds.hycom.org/thredds/dodsC/GLBv0.08/expt_57.7';
+date = '28-Aug-2017 00:00:00';  
+variable = 'salinity;                
+region = [-5.0, 45.0 ,160,-150 ];      
+depth = -150;                                                   
+model_simple_plot(model,source,date,variable,region,depth);
 
-% Load Argo data from west of New Zealand:
+% Click stations on the plot to create a coordinate list:
 
-region = [-60.0 -50.0 150.0 160.0]; %  Search region [-90 90 -180 180]
-start_date = '01-Nov-2015 00:00:00';
-end_date = '01-Jan-2017 00:00:00';
-variable_list = {'TEMP_ADJUSTED','PSAL_ADJUSTED'};
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list);
-
-% Make plots:
-
-general_profiles(argo,variable,'depth')
-general_map(argo,bathymetry_dir,'2Dcontour')
+[xcoords,ycoords] = transect_select(10); % click desired transect on the figure, densify selection by 10x
 
 ```
-<img src="https://user-images.githubusercontent.com/24570061/88301724-fd1dab80-ccd2-11ea-9ea7-7badf1424865.png" width="600">
-<img src="https://user-images.githubusercontent.com/24570061/88301788-11fa3f00-ccd3-11ea-9cdf-1622f701bfe9.png" width="600">
+<img src="https://user-images.githubusercontent.com/24570061/88406388-9f569580-cd9e-11ea-9871-e4d55941d7c4.png" width="500">
 
-[Back](https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1)
+```Matlab
+
+% Build a uniform struct from the coordinates:
+
+variable_list = {'water_temp','salinity'}; % 'water_u' 'water_v' 'water_temp' 'salinity'
+[hycom] =  model_build_profiles(source,date,variable_list,xcoords,ycoords);
+
[…]
