### 2020-07-24 da4378c
Update model_simple_plot.md
