+++ docs/model_simple_plot.md
@@ -1,41 +1,39 @@
-### argo_build
+### model_simple_plot
 
 #### Syntax
 
 ```Matlab
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)
+[data,lat,lon] = model_simple_plot(model,source,date,variable,region,depth)
+[data,lat,lon] = model_simple_plot(model,source,date,variable,region,depth,arrows)
 ```
 #### Description
 
-``[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)`` searches pathway ``argo_dir`` for profiles meeting the search criteria ``region``, ``start_date``, and ``end_date``. Profiles are loaded into the struct array ``argo`` with all variables specified in ``variable_list``. Variables PLATFORM_NUMBER, LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically. Files containing matching profiles are listed in ``matching_files``.
+``[data,lat,lon] = model_simple_plot(model,source,date,variable,region,depth)`` plots one ``depth`` level of HYCOM or Operational Mercator GLOBAL_ANALYSIS_FORECAST_PHY_001_024. ``variable`` specifies the parameter to be plotted and ``region`` is the rectangular region to be plotted. ``model='hycom'`` or ``model='mercator'`` specifies the model used. ``source`` is the url or local path of the relevant dataset. ``data``, ``lat``, and ``lon`` from the plotted layer are available outputs.
 
+``[data,lat,lon] = model_simple_plot(model,source,date,variable,region,depth,arrows)`` adds directional arrows if it is a velocity magnitude plot. ``arrows=1`` is on, ``arrows=0`` is off.
 
 #### Example 1
 
 
 ```Matlab
 
-% Get variable information:
+% Setup nctoolbox: 
 
-argo_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/argo/*profiles*.nc';
-netcdf_info(argo_dir);
+setup_nctoolbox
 
-% Load Argo data from west of New Zealand:
+% Plot temperature at the depth level closes to 150m:
 
-region = [-60.0 -50.0 150.0 160.0]; %  Search region [-90 90 -180 180]
-start_date = '01-Nov-2015 00:00:00';
-end_date = '01-Jan-2017 00:00:00';
-variable_list = {'TEMP_ADJUSTED','PSAL_ADJUSTED'};
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list);
-
-% Make plots:
-
-general_profiles(argo,variable,'depth')
-general_map(argo,bathymetry_dir,'2Dcontour')
+model = 'mercator'; % 'hycom' 'mercator'
+source = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/mercator/global-analysis-forecast-phy-001-024_1593408360353.nc'; 
+date = '18-Mar-2020 00:00:00';   
+variable = 'thetao'; 
+region = [60.0, 70.0 ,-80, -60];      % [-90 90 -180 180]
+depth = -150;                
+arrows = 0;  
+model_simple_plot(model,source,date,variable,region,depth,arrows)
 
 ```
-<img src="https://user-images.githubusercontent.com/24570061/88301724-fd1dab80-ccd2-11ea-9ea7-7badf1424865.png" width="600">
-<img src="https://user-images.githubusercontent.com/24570061/88301788-11fa3f00-ccd3-11ea-9cdf-1622f701bfe9.png" width="600">
+<img src="https://user-images.githubusercontent.com/24570061/88408553-b8147a80-cda1-11ea-98bf-e1d9b45aa53c.png" width="600">
 
[…]
