### 2020-07-24 355dd6d
Update model_simple_plot.md
