### 2020-07-24 98094f6
Update model_simple_plot.md
