### 2020-07-24 bc4cae7
Update model_simple_plot.md
