### 2020-07-24 cf2ffa9
Update model_simple_plot.md
