### 2020-07-24 fce61ff
Update model_domain_plot.md
