### 2020-07-24 90c67f0
Update model_simple_plot.md
