+++ docs/model_build_profiles.md
@@ -1,41 +1,75 @@
-### argo_build
+### model_build_profiles
 
 #### Syntax
 
 ```Matlab
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)
+[model] = model_build_profiles(source,date,variable_list,xcoords,ycoords)
+[model] =  model_build_profiles(source,date,variable_list,xcoords,ycoords,zgrid)
 ```
 #### Description
 
-``[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)`` searches pathway ``argo_dir`` for profiles meeting the search criteria ``region``, ``start_date``, and ``end_date``. Profiles are loaded into the struct array ``argo`` with all variables specified in ``variable_list``. Variables PLATFORM_NUMBER, LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically. Files containing matching profiles are listed in ``matching_files``.
+``[model] = model_build_profiles(source,date,variable_list,xcoords,ycoords)`` builds a struct of profiles from HYCOM or Operational Mercator GLOBAL_ANALYSIS_FORECAST_PHY_001_024, pulling profiles nearest to coordinates specified by ``xcoords`` and ``ycoords``. Profiles are loaded into the struct array ``model`` with all variables specified in ``variable_list``.
 
+``[model] = model_build_profiles(source,date,variable_list,xcoords,ycoords,zgrid)`` depth-interpolates the profiles to a vertical grid of ``zgrid``, in meters. ``zgrid=2`` would produce profiles interpolated to 2 meter vertical grid.
 
+HYCOM variables: 
+```Matlab
+'water_u' 
+'water_v' 
+'water_temp' 
+'salinity' 
+```
+Mercator variables: 
+```Matlab
+'thetao' 
+'so' 
+'uo' 
+'vo' 
+```
+                     
 #### Example 1
 
 
 ```Matlab
 
-% Get variable information:
+% Setup nctoolbox: 
 
-argo_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/argo/*profiles*.nc';
-netcdf_info(argo_dir);
+setup_nctoolbox
 
-% Load Argo data from west of New Zealand:
+% Plot temperature at the depth level closest to 150m:
 
-region = [-60.0 -50.0 150.0 160.0]; %  Search region [-90 90 -180 180]
-start_date = '01-Nov-2015 00:00:00';
-end_date = '01-Jan-2017 00:00:00';
-variable_list = {'TEMP_ADJUSTED','PSAL_ADJUSTED'};
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list);
+model = 'mercator'; % 'hycom' 'mercator'
+source = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/mercator/global-analysis-forecast-phy-001-024_1593408360353.nc'; 
+date = '18-Mar-2020 00:00:00';   
+variable = 'thetao'; 
+region = [60.0, 70.0 ,-80, -60];      % [-90 90 -180 180]
+depth = -150;                
+arrows = 0;  
+model_simple_plot(model,source,date,variable,region,depth,arrows)
+
+% Click stations on the plot to create a coordinate list:
+
+[xcoords,ycoords] = transect_select(10); % click desired transect on the figure, densify selection by 10x 
+
+```
+<img src="https://user-images.githubusercontent.com/24570061/88411026-3c1c3180-cda5-11ea-81d3-d3b656315464.png" width="600">
+
+```Matlab
+
+% Build a uniform struct of profiles:
+
[…]
