### 2020-07-24 650bc0b
Update netcdf_info.md
