### 2020-07-24 3ddc891
Update netcdf_info.md
