+++ README.md
@@ -130,7 +130,7 @@ To take full advantage of ocean_data_tools, download [Gibbs-SeaWater (GSW) Ocean
 
 ### Miscellaneous utilities
 
-**netcdf_info** gets information about the first netcdf file in a path and saves it to a text file
+**[netcdf_info](docs/netcdf_info.md)** gets information about the first netcdf file in a path and saves it to a text file
 
 **region_select** creates coordinate list (which represents vertices of a polygon region) by clicking stations on a plot
 
