### 2020-07-24 lnferris/ocean_data_tools#2 @lnferris
@kthyng do I close issues once fixed or do I let reviewers decide whether to close them? Thanks!
