### 2020-07-24 3b3efe0
Update general_region_subset.md
