### 2020-07-24 f02fe41
Update general_depth_subset.md
