### 2020-07-24 1127dfb
Update general_depth_subset.md
