### 2020-07-24 0ca9e60
Update general_map.md
