### 2020-07-24 2dab57e
Update general_remove_duplicates.md
