### 2020-07-24 b220c92
Update general_map.md
