### 2020-07-24 fcbd57f
Update general_map.md
