+++ docs/general_map.md
@@ -1,4 +1,4 @@
-### argo_build
+### general_map
 
 #### Syntax
 
