### 2020-07-24 12344c2
Update general_map.md
