### 2020-07-24 4ec7bb7
Update general_map.md
