+++ README.md
@@ -70,7 +70,7 @@ To take full advantage of ocean_data_tools, download [Gibbs-SeaWater (GSW) Ocean
 
 **[general_depth_subset](docs/general_depth_subset.md)** subsets a uniform struct by depth
 
-**general_map** plots coordinate locations in a uniform struct, with optional bathymetry contours
+**[general_map](docs/general_map.md)** plots coordinate locations in a uniform struct, with optional bathymetry contours
 
 **general_profiles** plots vertical profiles in a uniform struct
 
