### 2020-07-24 578c778
Update general_section.md
