+++ docs/general_section.md
@@ -1,13 +1,19 @@
-### argo_build
+### general_section
 
 #### Syntax
 
 ```Matlab
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)
+general_section(object,variable,xref,zref)
+general_section(object,variable,xref,zref,interpolate)
+general_section(object,variable,xref,zref,interpolate,contours)
 ```
 #### Description
 
-``[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)`` searches pathway ``argo_dir`` for profiles meeting the search criteria ``region``, ``start_date``, and ``end_date``. Profiles are loaded into the struct array ``argo`` with all variables specified in ``variable_list``. Variables PLATFORM_NUMBER, LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically. Files containing matching profiles are listed in ``matching_files``.
+``general_section(object,variable,xref,zref)`` creates a section plot from ``object``; where ``object`` is a struct created by any of the ``_build`` functions in ocean_data_tools (e.g. ``argo``, ``cruise``, ``hycom``, ``mercator``, ``woa``, ``wod``). The color field is specified by ``variable``. ``xref`` and ``zref`` specify fields to use for the x-axis and z-axis.
+
+``general_section(object,variable,xref,zref,interpolate)`` interpolates the plot using the MATLAB ``shading`` function. ``interpolate=1`` for on, ``interpolate=0`` for off.
+
+``general_section(object,variable,xref,zref,interpolate,contours)`` adds contours to the section plot. ``contours=1`` for on, ``contours=0`` for off.
 
 
 #### Example 1
@@ -15,27 +21,42 @@
 
 ```Matlab
 
-% Get variable information:
+% Setup nctoolbox:
 
-argo_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/argo/*profiles*.nc';
-netcdf_info(argo_dir);
+setup_nctoolbox
 
-% Load Argo data from west of New Zealand:
+% Built a uniform struct from MOCHA climatology:
 
-region = [-60.0 -50.0 150.0 160.0]; %  Search region [-90 90 -180 180]
-start_date = '01-Nov-2015 00:00:00';
-end_date = '01-Jan-2017 00:00:00';
-variable_list = {'TEMP_ADJUSTED','PSAL_ADJUSTED'};
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list);
-
-% Make plots:
-
-general_profiles(argo,variable,'depth')
-general_map(argo,bathymetry_dir,'2Dcontour')
+month = 10; % Month (1 through 12).
+depth = 0;
+variable = 'temperature'; %  'temperature' 'salinity'
+region = [34 42  -80 -70]; % [30 48 -80 -58]
+mocha_simple_plot(month,depth,variable,region)
+[xcoords,ycoords] = transect_select(10); % click desired transect on the figure, densify selection by 10x 
 
 ```
-<img src="https://user-images.githubusercontent.com/24570061/88301724-fd1dab80-ccd2-11ea-9ea7-7badf1424865.png" width="600">
-<img src="https://user-images.githubusercontent.com/24570061/88301788-11fa3f00-ccd3-11ea-9cdf-1622f701bfe9.png" width="600">
 
-[Back](https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1)
+<img src="https://user-images.githubusercontent.com/24570061/88334226-73d09e00-ccff-11ea-867d-860d64744dc0.png" width="600">
+
+```Matlab
+
+% Build a uniform struct of profiles:
+
[…]
