### 2020-07-24 434a70b
Update general_section.md
