+++ docs/general_section.md
@@ -33,17 +33,12 @@ variable = 'temperature'; %  'temperature' 'salinity'
 region = [34 42  -80 -70]; % [30 48 -80 -58]
 mocha_simple_plot(month,depth,variable,region)
 [xcoords,ycoords] = transect_select(10); % click desired transect on the figure, densify selection by 10x 
-
+zgrid = 1; % vertical grid for linear interpolation in meters
+[mocha] = mocha_build_profiles(month,xcoords,ycoords,zgrid); % zgrid optional, no interpolation if unspecified
 ```
-
 <img src="https://user-images.githubusercontent.com/24570061/88334226-73d09e00-ccff-11ea-867d-860d64744dc0.png" width="600">
 
 ```Matlab
-
-% Build a uniform struct of profiles:
-
-zgrid = 1; % vertical grid for linear interpolation in meters
-[mocha] = mocha_build_profiles(month,xcoords,ycoords,zgrid); % zgrid optional, no interpolation if unspecified
 
 % Make a temperature section:
 
