### 2020-07-24 d11464b
Update general_remove_duplicates.md
