### 2020-07-24 5816fac
Update bathymetry_extract.md
