### 2020-07-24 1d7b52a
Update bathymetry_section.md
