### 2020-07-24 8e0640e
Update bathymetry_section.md
