### 2020-07-24 1a40561
Update bathymetry_chord.md
