### 2020-07-24 c31d8ce
Update bathymetry_chord.md
