### 2020-07-24 b2b8fe5
Update bathymetry_chord.md
