### 2020-07-24 1ae36f3
Update bathymetry_plot.md
