+++ docs/bathymetry_plot.md
@@ -1,41 +1,38 @@
-### argo_build
+### bathymetry_plot
 
 #### Syntax
 
 ```Matlab
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)
+bathymetry_plot(bathymetry_dir,region,ptype)
 ```
 #### Description
 
-``[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)`` searches pathway ``argo_dir`` for profiles meeting the search criteria ``region``, ``start_date``, and ``end_date``. Profiles are loaded into the struct array ``argo`` with all variables specified in ``variable_list``. Variables PLATFORM_NUMBER, LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically. Files containing matching profiles are listed in ``matching_files``.
-
-
+``bathymetry_plot(bathymetry_dir,region,ptype)`` extracts Smith & Sandwell Global Topography in path ``bathymetry_dir`` for use with a 2D (latitude vs. longitude) or 3D (latitude vs. longitude vs. depth) plot. The region extracted is given by ``region``, maximum latitudes and longitudes are [-80.738 80.738 -180 180], and should be in order [S N W E]. ``type = '2Dscatter'`` or ``'2Dcontour'`` or ``'3Dsurf'`` specifies the plot type.
+                     
 #### Example 1
-
 
 ```Matlab
 
-% Get variable information:
+% Setup nctoolbox: 
 
-argo_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/argo/*profiles*.nc';
-netcdf_info(argo_dir);
+setup_nctoolbox
 
-% Load Argo data from west of New Zealand:
+% Plot a 3-D velocity domain from Operational Mercator:
 
-region = [-60.0 -50.0 150.0 160.0]; %  Search region [-90 90 -180 180]
-start_date = '01-Nov-2015 00:00:00';
-end_date = '01-Jan-2017 00:00:00';
-variable_list = {'TEMP_ADJUSTED','PSAL_ADJUSTED'};
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list);
+model = 'mercator'; % 'hycom' 'mercator'
+source = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/mercator/global-analysis-forecast-phy-001-024_1593408360353.nc'; 
+date = '18-Mar-2020 00:00:00';   
+variable = 'thetao'; 
+region = [60.0, 70.0 ,-80, -60];      % [-90 90 -180 180]
+variable = 'velocity'; 
+model_domain_plot(model,source,date,variable,region)
 
-% Make plots:
+% Add bathymetry:
 
-general_profiles(argo,variable,'depth')
-general_map(argo,bathymetry_dir,'2Dcontour')
+bathymetry_plot(bathymetry_dir,region,'3Dsurf')
+caxis([0 1])
 
 ```
-<img src="https://user-images.githubusercontent.com/24570061/88301724-fd1dab80-ccd2-11ea-9ea7-7badf1424865.png" width="600">
-<img src="https://user-images.githubusercontent.com/24570061/88301788-11fa3f00-ccd3-11ea-9cdf-1622f701bfe9.png" width="600">
-
-[Back](https://github.com/lnferris/ocean_data_tools#building-uniform-structs-from-data-sources-1)
+<img src="https://user-images.githubusercontent.com/24570061/88409944-ab912180-cda3-11ea-84bc-f848a4f795bc.png" width="700">
+[Back](https://github.com/lnferris/ocean_data_tools#adding-bathymetry-to-existing-plots-1)
 
