### 2020-07-24 3c61c7a
Update bathymetry_plot.md
