### 2020-07-24 lnferris/ocean_data_tools#3 @lnferris
I have added linked markdown files for all functions.
