+++ ocean_data_tools/bathymetry_section.m
@@ -1,12 +1,12 @@
 
-function [bath_section,lon_section,lat_section] = bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled)
+function [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled)
 % bathymetry_section adds global seafloor topography (Smith & Sandwell, 1997) to an existing section
 % plot from points nearest to specified coordinates
 % 
 %% Syntax
 % 
-%  [bath_section,lon_section,lat_section] = bathymetry_section(bathymetry_dir,xcoords,ycoords,xref)
-%  [bath_section,lon_section,lat_section] = bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled)
+%  [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathymetry_dir,xcoords,ycoords,xref)
+%  [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled)
 % 
 %% Description 
 % 
@@ -15,8 +15,10 @@ function [bath_section,lon_section,lat_section] = bathymetry_section(bathymetry_
 % Smith & Sandwell Global Topography in path bathymetry_dir for use with a
 % section plot. Points are extracted nearest to each coordinate specified
 % by xcoords (longitude) and ycoords (latitude). The bathymetry section is
-% plotted with xref = 'lon' or xref = 'lat' as the x-axis variable. The
-% extracted data is output bath_section, lon_section, and lat_section.
+% plotted with xref; where xref = 'lon', 'lat', or a time vector of length(xcoords). The
+% extracted data is output bath_section, lon_section, and lat_section; 
+% output vectors are sorted by the selected reference axis (longitude,
+% latitude, or time).
 %  
 % [bath_section,lon_section,lat_section] =
 % bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled) allows the
@@ -43,37 +45,50 @@ function [bath_section,lon_section,lat_section] = bathymetry_section(bathymetry_
 % figure
 % bathymetry_section(bathymetry_dir,xcoords,ycoords,xref)
 %
+%% Example 3
+% Plot bathymetry nearest to a list of coordinates. Use a time as the x-axis:
+% 
+% xref = [737009 737010 737011 737012 737013]; 
+% xcoords = [60 60.1 60.4 60.2 59.9]; 
+% ycoords = [10 20.1 15.0 16.1 13.7]; 
+% figure
+% bathymetry_section(bathymetry_dir,xcoords,ycoords,xref)
+%
 %% Citation Info 
 % github.com/lnferris/ocean_data_tools
-% Jun 2020; Last revision: 30-Jun-2020
+% Jun 2020; Last revision: 26-Jul-2020
 % 
 % See also general_section and bathymetry_chord.
-
 
 if nargin < 5
     filled = 0;
 end
 
-% Auto-select region.
+% Auto-select region and load bathymetry data.
 region = [min(ycoords)-1 max(ycoords)+1 min(xcoords)-1 max(xcoords)+1];
-
-% Load bathymetry data.
 [bath,lat,lon] = bathymetry_extract(bathymetry_dir,region);
 
-% deal with inputs other than [-90 90 -180 180] e.g  [-90 90 20 200] 
+% Deal with inputs other than [-90 90 -180 180] e.g  [-90 90 20 200] 
 region(region>180) = region(region>180)- 360;
 region(region<-180) = region(region<-180)+360;
 
[…]
