+++ ocean_data_tools/bathymetry_section.m
@@ -1,5 +1,5 @@
 
-function [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled)
+function [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled,maxdistance)
 % bathymetry_section adds global seafloor topography (Smith & Sandwell, 1997) to an existing section
 % plot from points nearest to specified coordinates
 % 
@@ -7,6 +7,7 @@ function [bath_section,lon_section,lat_section,time_section] = bathymetry_sectio
 % 
 %  [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathymetry_dir,xcoords,ycoords,xref)
 %  [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled)
+%  [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled,maxdistance)
 % 
 %% Description 
 % 
@@ -24,7 +25,13 @@ function [bath_section,lon_section,lat_section,time_section] = bathymetry_sectio
 % bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled) allows the
 % bathymetry to be filled in black down to the x-axis (instead of a simple line).
 % Set filled=1 to turn on, filled=0 to turn off.
-% 
+%
+% [bath_section,lon_section,lat_section] =
+% bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled,maxdistance)
+% does not pull values where xcoords and ycoords are not within maxdistance (degrees) of a Global Topography
+% value. maxdistance=0.05 would pull no bathymetry at times/places not within
+% 0.05 diagonal degrees of an available Global Topography value.
+%
 %% Example 1
 % Add bathymetry to a temperature section plot from the list of coordinates
 % stored in struct cruise, filling in bathymetry. Use longitude as the x-axis:
@@ -54,6 +61,19 @@ function [bath_section,lon_section,lat_section,time_section] = bathymetry_sectio
 % figure
 % bathymetry_section(bathymetry_dir,xcoords,ycoords,xref)
 %
+%% Example 4
+% Plot bathymetry nearest to a list of coordinates, using time as the
+% x-axis and shading bathymetry. Only return bathymetry values for timesteps within 0.007 degrees
+% of a bathymetry node:
+% 
+% xref = [737009 737010 737011 737012 737013]; 
+% xcoords = [60 60.1 60.4 60.2 59.9]; 
+% ycoords = [10 20.1 15.0 16.1 13.7]; 
+% filled = 1;
+% maxdistance = 0.007;
+% figure
+% bathymetry_section(bathymetry_dir,xcoords,ycoords,xref,filled,maxdistance)
+%
 %% Citation Info 
 % github.com/lnferris/ocean_data_tools
 % Jun 2020; Last revision: 26-Jul-2020
@@ -79,6 +99,7 @@ if region(3) > region(4) && region(3) < 0 && region(4)+360>180
     bath = bath(lon_inds,:);  
 end
 
+% Get nearest bathymetry values.
 N = length(xcoords);
 if length(xref)==N 
     time_section = xref;
@@ -89,13 +110,24 @@ bath_section = NaN(1,N);
 lon_section = NaN(1,N);
 lat_section = NaN(1,N);
 for i = 1:N
-    [lon_ind,~] = near(lon,xcoords(i));
-    [lat_ind,~] = near(lat,ycoords(i));
-    bath_section(i) = bath(lon_ind,lat_ind);
[…]
