### 2020-07-28 710e9ec
Update bathymetry_section.m
