### 2020-07-28 919002c
Update bathymetry_section.md
