### 2020-08-01 97b786c
Update whp_cruise_build.md
