### 2020-08-01 e33a492
Update whp_cruise_build.md
