### 2020-08-01 0fdc2a7
Update whp_cruise_build.md
