### 2020-08-01 37a69e1
Update whp_cruise_build.md
