### 2020-08-11 #2497 @kthyng
Hi @kakearney! @chadagreene suggested that you would be a good reviewer for this package. Are you interested in reviewing this for JOSS? The review process is of the software itself and a short paper, and since this is already the review issue on github, you can see the checklist-based approach above. You can also get more information here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
I would hope for the review in the next 3-4 weeks if possible since this review time period has already started, but I can be flexible if needed.
