### 2020-08-11 #2497 @kthyng
@kakearney awesome! I am going to do a reviewer switcharoo on this issue above. I will invite you as a reviewer after a few steps and let me know if you don't have permission to edit the checklist after you accept the invitation (which should be emailed to you).
