### 2020-08-12 #2497 @castelao
Hi @kthyng , how do I update my review checklist? Was I supposed to have editing permissions on the very first post?
