### 2020-08-27 #2497 @kakearney
I’ll be submitting mine this afternoon... just compiling all my notes into
issues and a summary now.

On Thu, Aug 27, 2020 at 5:33 AM Kristen Thyng <notifications@github.com>
wrote:

>
>
> Hi @castelao <https://github.com/castelao> and @kakearney
> <https://github.com/kakearney>! Do you have a timeline in mind for when
> you could work on your reviews? Thanks!
>
>
>
>
