### 2020-08-27 lnferris/ocean_data_tools#11 @kakearney
Bathymetry data reloaded for plots
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

Bathymetry data is reloaded from file every time it needs to be plotted; this is a very inefficient way of working with this data, especially across larger regions.  It would be better if you separated data extraction from data plotting, allowing reuse of the same data in multiple plots.

Separating the two processes would also allow users to use a different bathymetry dataset; there's no particular reason to favor this one dataset, given that the bathymetry data is typically used only as a reference for plotting. Higher- or lower-resolution bathymetry products may be more useful for a given user.  This is particularly true for basin-scale regions, where 1-minute resolution is overkill.

Also, note that Matlab's Mapping Toolbox has built-in support for the Smith & Sandwell bathymetry data, via the `satbath` function, but `satbath` expects data in the .img format, while ocean\_data\_tools expects it in netCDF format.  Allowing more flexibility regarding bathymetry data can allow users like myself who already have the .img data downloaded to avoid unecessary replication. (I will note that the satbath function is designed around the older 2-minute dataset and is not compatible with newer versions due to its hard-coding of expected matrix dimensions, so there are good reasons to prefer bathymetry\_extract over satbath in some use cases... but for many users the lower-resolution data is sufficient.)
