### 2020-08-27 lnferris/ocean_data_tools#12 @kakearney
Remove eval
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

The use of `eval` is not recommended Matlab syntax.  Please replace with dynamic field names (I found similar use in argo\_profiles.m, general\_depth\_subset.m, general\_profiles.m, general\_section.m):

```matlab
cvar = eval(['argo.' variable]); % bad
cvar = argo.(variable); % good
```
