### 2020-08-27 lnferris/ocean_data_tools#13 @kakearney
Revise or remove more_colors
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

The more\_colors function should be eliminated.  It changes the user's default color order at the Matlab root level.  This is very undesireable behavior; functions should never alter user preferences like this, especially without warning (this function is undocumented and called from within other plotting functions).  There are far better ways to change the color of plotted lines, e.g. setting colors of specific line objects manually via `set()`, altering the color order of a single specific axis, etc.
