### 2020-08-27 lnferris/ocean_data_tools#14 @kakearney
Plotting function customization
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

Currently, many of the plotting functions are very heavy-handed.  They do things like create new figures and axes, change the hold state of existing axes, update colormaps and color order, set ticks and gridlines, etc.  I would reconsider whether many of these changes are necessary.  In particular, consider whether new axes need to be created, or if it makes sense to allow the user to target an existing axis.  If using an existing axis, consider whether aesthetic changes are necessary, or if they may unintentionally alter prexisting plotted objects.  At minimum, the documentation for the plotting functions should be updated to inform users any changes that will be applied.

Also, it's nice to have the option to customize aesthetic properties, either on creation (as input variables) or after the fact (by modifying existing graphics handles).  Currently, users must do some graphics handle excavation (via `findobj`) to make any of these types of changes.  I recommend updating all the plotting functions to save and return graphics handles associated with all newly-created graphics objects, including figures, axes, lines, contours, pcolors, etc.
