### 2020-08-27 lnferris/ocean_data_tools#15 @kakearney
Longitude wrapping
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

Functions that use a `region` box to filter data points need to be robust to varying longitude-wrapping conventions.  Right now, some (e.g. bathymetry\_extract) are, but others (e.g. argo\_build) assume that the user's input box will match the conventions of the underlying dataset.  Please check that all are either robust to any wrapping, or throw an informative error message for mismatches. At minimum, I suggest testing -180-180, 0-360, 25-385 (ocean-focused), and -25-335 (land-focused).
