### 2020-08-27 lnferris/ocean_data_tools#16 @kakearney
Revise or remove netcdf_info
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

I couldn't figure out the purpose of this function.  I can see that it's being used to store `ncdisp` output, but why?  No functions seem to access this info.  The only benefit I can see over simply calling `ncdisp` directly is that it accepts a file path rather than a single file name, but there is no checking to see if the first file found on the indicated search path is representative of all files matching this search pattern.  The function assumes that it will always have write permission in the current directory, which is not always true.  Finally, the function messes with the root diary state within a try/catch loop, which is asking for trouble.  For example, if you pass a nonexistent path to the function, the diary state is never toggled off.  Also, it fails to check the diary state before making any changes, so it will not properly reset things if the user did not have the diary off at the time `netcdf_info` was called.

I recommend seriously reconsidering the logic behind this function.  Users can already simply call `ncdisp` to get a quick peek at the netCDF header info.  Alternatively, `ncinfo` can return the info in a storeable format with far fewer opportunities for failure and unintended side effects.
