### 2020-08-27 lnferris/ocean_data_tools#17 @kakearney
argo_build confuses pressure and depth
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

Pressure is not the same thing as depth!  Currently, the `argo_build` function places the PRES_ADJUSTED variable data from an Argo float data in a field called `depth`, despite it actually holding sea level pressure (in decibars), which will be very misleading for users.  The mismatch between pressure (dbar) and depth (m) is small at the surface, but is far from insignificant at depth. Especially with Deep Argo floats coming online now, treating them equivalently could lead to major miscalculations.

If you want this function to return data along a depth axis, you need to properly convert, not just rename the field.  Otherwise, please rename the output field to properly reflect its contents.
