### 2020-08-27 lnferris/ocean_data_tools#18 @kakearney
Rename bathymetry_region
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

This function doesn't actually have anything to do with bathymetry, it simply returns a bounding region.  Please rename accordingly.
