### 2020-08-27 lnferris/ocean_data_tools#19 @kakearney
Bathymetry plotting along a transect
Both bathymetry\_section and bathymetry\_chord suffer from issues regarding how they order the output points when plotting against a transect.

For bathymetry\_section, there are two issues.  The first is that it uses the coordinates of the closest-matching-grid cell, rather than the input coordinates, along the x axis.  Unless the input line is parallel to the specified axis (i.e. an east/west line paired with lon, or a north/south line paired with lat), this may cause strong reduction in the plotted x-coordinate resolution relative to the input coordinates.  The points are also plotted in the order they appear in the input, rather than in increasing order with longitude and/or latitude, which can be an issue for any set of points whose x-coordinate dataset (x for xref='lon' or y for xref='lat') is not monotonically increasing.

For bathymetry\_chord, the output can get even messier.  The general idea behind this function appears to be to plot the bathymetry along a transect line.  However, it only does so in a very limited context: when the line is parallel to the specified x axis, and the specified `width` is approximately equal to or less than to the resolution of the bathymetry dataset.  Anything else results in a mess of lines doubling back on themselves, due to the way grid cell within the polygon are collected.

A demonstration:

```matlab
% A few lines near the Aleutian Islands

latlon = [...
       53.51       176.23
       53.74         -177
       55.055       179.68
       50.623       -179.8
       53.165       171.64
       51.687       178.25];
latlon(:,2) = wrapTo360(latlon(:,2));

latlim = minmax(latlon(:,1), 'expand');
lonlim = minmax(latlon(:,2), 'expand');

% Read bathymetry

bdir = '/Volumes/LacieKK2019/SmithSandwell/topo_20.1.nc'; % user-specific
[z,lt,ln] = bathymetry_extract(bdir, [latlim lonlim]);

cmap = get(0, 'defaultAxesColorOrder');

% Plot map of bathymetry

figure; axes;
pcolor(ln,lt,z');
shading flat;
colormap('gray');
hold on;
plot(latlon(1:2,2), latlon(1:2,1), 'color', cmap(1,:));
plot(latlon(3:4,2), latlon(3:4,1), 'color', cmap(2,:));
plot(latlon(5:6,2), latlon(5:6,1), 'color', cmap(3,:));
set(gcf, 'color', 'w');

d = {'lat', 'lon'};

% Plot transects

for ii = 1:3

    idx = (ii-1)*2+[1 2];
    [y, x] = interpm(latlon(idx,1), latlon(idx,2), 1/60);

    zline = interp2(ln,lt,z',x,y, 'nearest');

    h = plotgrid('size', [4 2], 'sp', 0.05);

    for id = 1:2
        % Wider chord
        axes(h.ax(1,id));
        bathymetry_chord(bdir, x(1), y(1), x(end), y(end), d{id},  0, 1/20);
        % Narrower chord
        axes(h.ax(2,id));
        bathymetry_chord(bdir, x(1), y(1), x(end), y(end), d{id},  0, 1/60);
        % Section
        axes(h.ax(3,id));
        bathymetry_section(bdir, x, y, d{id}, 0);
        % Section (my way)
        switch id
            case 1
                plot(h.ax(4,id), y, zline, 'color', 'k');
            case 2
                plot(h.ax(4,id), x, zline, 'color', 'k');
        end
    end

    set(findall(h.ax, 'type', 'line'), 'linewidth', 1, 'color', cmap(ii,:));
    set(h.ax(:,1), 'xlim', minmax(y));
    set(h.ax(:,2), 'xlim', minmax(x), 'yaxisloc', 'right');
    set(h.ax, 'ylim', minmax(zline));
    xlabel(h.ax(end,1), 'Latitude');
    xlabel(h.ax(end,2), 'Longitude');
    set(h.fig, 'color', 'w');
    set(h.ax(1:end-1,:), 'xticklabel', '');

    h.lbl = multitextloc(h.ax(:,1), {...
        'bathymetry_chord (width=1/20)'
        'bathymetry_chord (width=1/60)'
        'bathymetry_section'
        'the transect'}, 'northwestoutsideabove');
    set(h.lbl, 'interpreter', 'none');

end

```

![bathybug1](https://user-images.githubusercontent.com/5015092/91502977-358a4980-e87e-11ea-9f67-135c2d47f716.png)
![bathybug2](https://user-images.githubusercontent.com/5015092/91502982-391dd080-e87e-11ea-8c41-21fdcc55bda2.png)
![bathybug3](https://user-images.githubusercontent.com/5015092/91502992-40dd7500-e87e-11ea-9e54-4760ba966cd1.png)
![bathybug4](https://user-images.githubusercontent.com/5015092/91502998-4470fc00-e87e-11ea-89d2-55806c9ea487.png)

(Note: this code uses some of my personal toolbox: [plotgrid](https://github.com/kakearney/plotgrid-pkg), [minmax](https://github.com/kakearney/minmax-pkg), [multitextloc](https://github.com/kakearney/multitextloc-pkg))
