### 2020-08-27 lnferris/ocean_data_tools#21 @kakearney
Selection popup instructions
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

The `region_select` and `transect_select` functions could include more informative instructions in the popup boxes.  In `region_select`, let the users know they should draw a polygon.  For `transect_select`, the use of "stations" to describe both the clicked locations and the interpolated points confused me at first.
