### 2020-08-27 lnferris/ocean_data_tools#22 @kakearney
Remove external dependency on nctoolbox
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

This toolbox often relies on the java-based nctoolbox functions to read netCDF files, rather than using Matlab's native netCDF utilities (`ncread`, `ncinfo`, `netcdf` class methods, etc.).  The extra external dependency will be a barrier to use for some novice users.  Java dependencies also add an additional area for potential incompatiblity.
