### 2020-08-27 lnferris/ocean_data_tools#23 @kakearney
Unsuppressed command window output
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

A few functions seem to intentionally not suppress output (i.e. no terminal semi-colon on a code line) in order to print certain things to screen.  It's not clear to the me why these variables are being dumped to the command window.  For example, in `woa_domain_plot`, names of all variables in the file and the variable attributes of the selected-to-plot variable are dumped to the screen for no apparent reason.  I'd suggest using `disp` or `fprintf`, with enough explanatory text and formatting that the user understands the purpose and can clearly see that the printing is intentional (I assumed it was a mistake at first.)
