### 2020-08-27 lnferris/ocean_data_tools#24 @kakearney
Plot versus distance along transect
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

For the section-plotting tools, you offer the option to place either latitude or longitude along the x-axis.  Distance-along-transect would be another obvious candidate here, especially when section-plotting circuitous routes that may double back on themselves.
