### 2020-08-27 lnferris/ocean_data_tools#25 @kakearney
transect_select point density
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

The transect\_select function allows users to specify the density of points along each individual line segment.  It would be nice to allow users to specify density in terms of distance rather than number of points, or to specify number of points along the entire transect rather than along each segement.  As it is, a long segment followed by a short segment results in very uneven station-spacing along the entirety of the polyline.
