### 2020-08-27 lnferris/ocean_data_tools#27 @kakearney
GSW compatibility
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

I think the statements regarding the compatiblity of the ocean\_data\_tools structures with the Gibbs Seawater Toolbox are misleading.  Yes, certain fields from the ODT structures can be used as input to GSW functions, in the same way that any temperature, salinity, or pressure data can be used.  However, the structure format does not add any additional value to GSW.  In fact, it may cause problems, given that the the ODT functions do not import the variable meta-data that explain exactly what type of data are in each dataset, e.g. that the `TEMP_ADJUSTED` field from an Argo float is ITS-90 in-situ temperature (t in GSW syntax), while the Mercator model `thetao` field is potential temperature (pt in GSW syntax).

I excitedly hoped that ODT would provide some nice wrapper functions to help users naviagate the often-tricky nuances of applying the GSW equation-of-state calculations when starting from different temp/salinity/pressure combos, but it does not.  I would suggest removing this compatiblity claim unless some more concrete connection (either in terms of new functions, or extended documentation matching up ODT structure fields to the GSW expected inputs) is added.
