### 2020-08-27 lnferris/ocean_data_tools#28 @kakearney
Markdown documentation vs Matlab docs
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

You've written some nice markdown documentation for most of the functions, which can be viewed online (or with a markdown editor/viewer, but I would not assume users have access to the latter).  However, these are not compatible with Matlab's own documentation viewer.  It would be very beneficial to users if you added .html versions of these documentation files and made them accessible via Matlab's `doc` command (see https://www.mathworks.com/help/matlab/matlab_prog/display-custom-documentation.html).
