### 2020-08-27 lnferris/ocean_data_tools#30 @kakearney
Automated tests
This issue is part of the JOSS review (openjournals/joss-reviews#2497)

While explicit unit tests are not necessary, I do think each function needs to be subjected to a much wider variety of input tests to make sure it either runs properly or throws an informative error.
