### 2020-08-27 #2497 @kakearney
In my review checklist, I have checked off the places where I think this submission meets the requirements of JOSS.  Overall, I believe this is a nice toolbox that fills a need in the ocean data community.  However, I do have a few concerns about the current functionality of the toolbox.  I believe the issues I have raised should be able to be corrected without too much additional effort.

To test the toolbox, I worked through the demo examples, adding a few variations of my own to each function call.  The following issues are ones that I believe need to be addressed before this toolbox is published in JOSS:

- Input parsing and validation (lnferris/ocean_data_tools#10)
- Bathymetry data reloaded for plots (lnferris/ocean_data_tools#11)
- Remove eval (lnferris/ocean_data_tools#12)
- Revise or remove more\_colors (lnferris/ocean_data_tools#13)
- Plotting function customization (lnferris/ocean_data_tools#14)
- Longitude wrapping (lnferris/ocean_data_tools#15)
- Revise or remove netcdf\_info (lnferris/ocean_data_tools#16)
- argo\_build confuses pressure and depth (lnferris/ocean_data_tools#17)
- Rename bathymetry\_region (lnferris/ocean_data_tools#18)
- Bathymetry plotting along a transect (lnferris/ocean_data_tools#19)

I also sumbitted a number of enhancement requests as issues.  These are not things that I believe need to be addressed before publication, but would be nice if they could be addressed eventually:

- general\_depth\_subset z-axis (lnferris/ocean_data_tools#20)
- Selection popup instructions (lnferris/ocean_data_tools#21)
- Remove external dependency on nctoolbox (lnferris/ocean_data_tools#22)
- Unsuppressed command window output (lnferris/ocean_data_tools#23)
- Plot versus distance along transect (lnferris/ocean_data_tools#24)
- transect\_select point density (lnferris/ocean_data_tools#25)

Regarding documentation, this submission includes a fair amount of it.  However, given that the target audience is likely to consist at least in part of novice programmers, I believe it could be more comprehensive.  In particular, the in-file header documentation needs to explicitly describe the input requirements for each function.  My comments on documentation have been submitted as issues:

- Installation instructions (lnferris/ocean_data_tools#26)
- GSW compatibility (lnferris/ocean_data_tools#27)
- Markdown documentation vs Matlab docs (lnferris/ocean_data_tools#28)
- In-file documentation (lnferris/ocean_data_tools#29)
- Automated tests (lnferris/ocean_data_tools#30)
- Demo script hard-coded paths (lnferris/ocean_data_tools#31)

The overview paper provides a very brief summary of the toolbox's primary purpose. Given the target audience, I suggest providing a bit more context.  For example, the first paragraph introduces the "structs" that are the common currency of this toolbox.  I suggest defining this term, describing the format in more detail (e.g. what types of data are possible fields in these structures?), and explaining why this method of organizing data may be more user-friendly than the native data storage of underlying datasets.  Note that while "struct" is common shorthand for "structure array" among Matlab users, the Matlab documentation favors using the full name whenever discussing variables of this class, and I recommend this paper do likewise.  Also, the paper does not mention why Matlab was chosen as the language of choice for this toolbox.  I recommend highlighting that Matlab is already extensively used within the oceanographic community.

The final paragraph mentions that the toolbox "has already been used in scientific publications."  I scanned the Crear et al., 2020 paper (Bemis et al., 2020 requires registration to view, and I opted out), and it doesn't appear that the toolbox data structures or plots are used directly in the paper.  I recommend changing the phrasing here to indicate that the toolbox was used for data exploration in support of these publications, rather than as the focal point of the studies.

Finally, as a generic ocean data exploration tool, I recognize that's its a bit difficult to state how this toolbox fits into the state of the field.  There are a number of semi-related ocean/climate Matlab toolboxes out there: for example, Chad Greene's Climate Data Toolbox, Rich Pawlowicz's various tools (tides, air-sea fluxes, mapping, etc.), Rich Signell's SEA-MAT tools, the Oceans Toolbox, the GSW Toolbox (and its deprecated but still commonly-used precessor, the seawater toolbox).  I'm not aware of any other publically-available, documentated, and designed-to-be-shared toolbox filling the same data exploration niche as this one.  Perhaps speak to this point somewhere in the paper?

To the authors, I hope the sudden flood of issues does not appear too negative!  I think this toolbox will serve the needs of a wide variety of users intimidated by the ever-growing variety of field- and model-derived oceanographic datasets. I look forward to seeing how this toolbox evolves in the future.

-Kelly Kearney
