### 2020-08-28 #2497 @lnferris
Thank you for your extremely detailed review, @kakearney!! I will get started on these edits as soon as I can. @kthyng, what is my timeline for making corrections?
