### 2020-08-31 #2497 @kthyng
@lnferris There isn't a specific deadline, but I would say if you plan to take a hiatus from this review issue for more than a few weeks, we can pause the issue in the meantime while you work in the background. If you want to work and reply here along the way for the various issues, then we don't need to pause and you can just work along. Does that make sense?
