### 2020-09-01 #2497 @castelao
@lnferris and @kthyng , I'm sorry for my slow response. I'll point my comments as issues in the project repository, and I'll make it clear what I consider critical from suggestions.

Impressive the review from @kakearney. Nice job!
