### 2020-09-01 lnferris/ocean_data_tools#32 @castelao
Take advantage of ERDDAP
This issue is part of the JOSS review ([openjournals/joss-reviews#2497](https://github.com/openjournals/joss-reviews/issues/2497))

A suggestion. Take advantage of [Ifremer's ERDDAP server](https://www.ifremer.fr/erddap/tabledap/ArgoFloats.html) to select profiles. It is useful to operate with a local directory but for some situations, it could be convenient to select profiles directly from the ERDDAP server instead. [Argopy](https://pypi.org/project/argopy/) could be a nice reference on how to implement that.

That would require some work, so maybe a good start is to just point the ERDDAP server in the documentation.
