### 2020-09-01 lnferris/ocean_data_tools#33 @castelao
DOI for each release
This issue is part of the JOSS review ([openjournals/joss-reviews#2497](https://github.com/openjournals/joss-reviews/issues/2497))

A suggestion. Since you are already using Zenodo, it might be useful to have a DOI for each version released, so when ocean_data_tools is used it can be pointed to the specific version. If you adopt that, [you can automate this process with Zenodo](https://guides.github.com/activities/citable-code/), and in that case, you can link the Zenodo DOI with your future JOSS paper (once accepted), as well as the DOIs of the datasets that you use, such as Argo dataset or WOA, and finally your ORCID. Here is an example of [how to configure the Zenodo DOI record](https://github.com/castelao/inception).

In my experience, the scientific community is still not used to cite software, so it is best for you to explicitly suggest to them what to cite when ocean_data_tools is used, and the JOSS DOI will probably be your best option.
