### 2020-09-01 lnferris/ocean_data_tools#34 @castelao
Documentation
This issue is part of the JOSS review ([openjournals/joss-reviews#2497](https://github.com/openjournals/joss-reviews/issues/2497))

Required. @kakearney made some good comments on the documentation #26, #28 & #29, and I think that those should not be simply suggestions, but should be addressed for the JOSS publication.

There is already a lot of information, but I found the documentation a little confusing. My first academic code packages were quite visionary (I think) but almost no one used because of my poor documentation at that time. There is really no limit on how much time we can invest in the documentation, but I think this is your best chance to create a good core. It will be hard to make time for that later. I reinforce the suggestion from @kakearney to improve the installation procedure, otherwise people get frustrated and give up without even trying your package. Experienced users probably already have their own system to fetch data. I think that a good fraction of your future users will be people learning, so abuse of clear and complete information, step by step.

Suggestion: Have you considered using the wiki resource from Github? It might be a convenient way to organize and maintain your documentation. It should be easy to move from /docs into the wiki, and you can clone the wiki itself. This is your call, I'm fine if you would rather use the current /docs. As your package evolves your README will turn into a huge file.

Another suggestion, I usually prefer a clean README with essential information about the package. If someone lands here, it should have just information enough to decide, "I'm the right place, let me learn more about this. Where are the examples?", or "this is not for me, let me keep searching for what I need".
