### 2020-09-04 #2497 @lnferris
@castelao and @kakearney thank you both for your comments! @kthyng I can start working on these issues now.
