### 2020-09-05 638b430
remove more_colors()
