### 2020-09-05 lnferris/ocean_data_tools#13 @lnferris
Thanks for the explanation. I have eliminated more_colors().
