### 2020-09-05 094fb1b
remove netcdf_info()
