+++ ocean_data_tools/whp_cruise_build.m
@@ -24,7 +24,8 @@ function [cruise] = whp_cruise_build(ctdo_dir,uv_dir,wvke_dir,variable_list)
 % uv_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/whp_cruise/uv/*.nc';
 % wvke_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/whp_cruise/wvke/';
 % 
-% netcdf_info(ctdo_dir) % Inspect netcdf to choose variable_list
+% listing = dir(ctdo_dir); % Peek at netCDF header info to inform choice of variable_list.
+% ncdisp([listing(1).folder '/' listing(1).name])
 % variable_list = {'salinity','temperature','oxygen'};
 % [cruise] = whp_cruise_build(ctdo_dir,uv_dir,wvke_dir,variable_list); % Use a dummy path (e.g. uv_dir ='null') if missing data. 
 %
