### 2020-09-05 0a69e7e
Update argo_platform_map.md
