### 2020-09-05 f72b27a
Update general_map.md
