### 2020-09-05 ffb00d6
Delete netcdf_info.md
