+++ docs/netcdf_info.md
@@ -1,22 +1,0 @@
-### netcdf_info
-
-#### Syntax
-
-```Matlab
-netcdf_info(nc_dir)
-```
-#### Description
-
-``netcdf_info(nc_dir)`` gets and information about the first file in path ``nc_dir`` and saves it to the working directory, creating or overwriting the text file "netcdf_info-ocean_data_tools"
-
-#### Example 1
-
-```Matlab
-% Get information about Argo netcdf files:
-
-argo_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/argo/*profiles*.nc';
-netcdf_info(argo_dir);
-```
-<img src="https://user-images.githubusercontent.com/24570061/88412188-2a3b8e00-cda7-11ea-921c-fa4c112adacf.png" width="800">
-
-[Back](https://github.com/lnferris/ocean_data_tools#miscellaneous-utilities-1)
