### 2020-09-05 24e5433
Update whp_cruise_build.md
