### 2020-09-05 58f73d4
Update wod_build.md
