### 2020-09-05 lnferris/ocean_data_tools#16 @lnferris
I have removed netcdf_info.m and replaced calls to this function (in demos and documentation) with an example of how to use ncdisp to peek at a file's header info.
