### 2020-09-05 #2497 @lnferris
I have fixed and closed "Revise or remove netcdf_info (lnferris/ocean_data_tools#16)".
