+++ ocean_data_tools/bathymetry_extract.m
@@ -25,14 +25,14 @@ function [bath,lat,lon] = bathymetry_extract(bathymetry_dir,region)
 % Extract relevant bathymetry around struct argo:
 %
 % bathymetry_dir = '/Users/lnferris/Documents/data/bathymetry/topo_20.1.nc';
-% region = bathymetry_region(argo);
+% region = bounding_region(argo);
 % [bath,lat,lon] = bathymetry_extract(bathymetry_dir,region);
 %
 %% Citation Info 
 % github.com/lnferris/ocean_data_tools
 % Jun 2020; Last revision: 30-Jun-2020
 % 
-% See also bathymetry_region, bathymetry_section, bathymetry_chord, and bathymetry_plot.
+% See also bounding_region, bathymetry_section, bathymetry_chord, and bathymetry_plot.
 
 % deal with inputs other than [-90 90 -180 180] e.g  [-90 90 20 200] 
 region(region>180) = region(region>180)- 360;
