### 2020-09-05 f0cf797
Update bounding_region.md
