### 2020-09-05 a40adff
Update argo_platform_map.md
