### 2020-09-05 d153dce
Update argo_platform_subset.md
