### 2020-09-05 lnferris/ocean_data_tools#18 @lnferris
I have renamed bathymetry_region to bounding_region.
