### 2020-09-05 #2497 @lnferris
I have fixed and closed "Rename bathymetry_region (lnferris/ocean_data_tools#18)".
