### 2020-09-05 #2497 @lnferris
I have fixed and closed "argo_build confuses pressure and depth (lnferris/ocean_data_tools#17)".
