### 2020-09-06 lnferris/ocean_data_tools#12 @lnferris
All uses of eval are replaced with dynamic field names. Thanks!
