### 2020-09-06 #2497 @lnferris
I have fixed and closed "Remove eval (lnferris/ocean_data_tools#12)".
