### 2020-09-06 b692b0e
demo.m automatic paths
