### 2020-09-06 lnferris/ocean_data_tools#31 @lnferris
Good suggestion. I have fixed it so pathways to test data are generated relative to demo.m itself. To help the new user transition to specifying paths to their own data, there is a subsequent block in which the user can experiment with hard-coding pathways.
