### 2020-09-06 #2497 @lnferris
I have fixed and closed "Demo script hard-coded paths (lnferris/ocean_data_tools#31)".
