### 2020-09-07 e105f32
Update bathymetry_extract.md
