### 2020-09-07 0fb291e
Update bathymetry_extract.md
