+++ docs/bathymetry_extract.md
@@ -3,11 +3,11 @@
 #### Syntax
 
 ```Matlab
-[bathy]=  bathymetry_extract(bathymetry_dir,region)
+[bathy] =  bathymetry_extract(bathymetry_dir,region)
 ```
 #### Description
 
-``[bathy]=  bathymetry_extract(bathymetry_dir,region)`` extracts Smith & Sandwell Global Topography in path ``bathymetry_dir`` over the specified rectangular ``region``. Struct ``bathy`` has fields ``z``, ``lat``, and ``lon``. This function is called by functions ``bathymetry_section``, ``bathymetry_chord``, and ``bathymetry_plot``.
+``[bathy] =  bathymetry_extract(bathymetry_dir,region)`` extracts Smith & Sandwell Global Topography in path ``bathymetry_dir`` over the specified rectangular ``region``. Struct ``bathy`` has fields ``z``, ``lat``, and ``lon``. This function is called by functions ``bathymetry_section``, ``bathymetry_chord``, and ``bathymetry_plot``.
 
 #### Example 1
 
