### 2020-09-07 b9da406
Update bathymetry_extract.md
