### 2020-09-07 6a2120b
Update bathymetry_section.md
