### 2020-09-07 f324126
avoid bathymetry reloading
