### 2020-09-07 3b2d597
Update argo_platform_map.md
