+++ docs/argo_platform_map.md
@@ -35,7 +35,7 @@ variable_list = {'TEMP_ADJUSTED','PSAL_ADJUSTED'};
 
 annotate = 1; 
 argo_platform_map(argo,annotate) % annotate optional,  1=on 0=off
-bathymetry_plot(bathymetry_dir,bounding_region(argo),'2Dcontour')
+bathymetry_plot(bathymetry_extract(bathymetry_dir,bounding_region(argo)),'2Dcontour')
 
 ```
 <img src="https://user-images.githubusercontent.com/24570061/88316847-6955da80-cce6-11ea-8bb0-d9d0523a3a29.png" width="700">
