### 2020-09-07 563d1bf
Update argo_platform_subset.md
