### 2020-09-07 b92da7d
Update bathymetry_plot.md
