### 2020-09-07 lnferris/ocean_data_tools#11 @lnferris
I have separated loading the data (using bathymetry_extract) from plotting the data (using bathymetry_plot and bathymetry_section).
