### 2020-09-07 #2497 @lnferris
I have fixed and closed "Bathymetry data reloaded for plots (lnferris/ocean_data_tools#11)".
