### 2020-09-07 b0a4531
Delete bathymetry_chord.md
