### 2020-09-07 aa93c03
clarify instructions
