### 2020-09-07 lnferris/ocean_data_tools#21 @lnferris
Added clarification. Thanks!
