### 2020-09-07 #2497 @lnferris
I have fixed and closed "Selection popup instructions (lnferris/ocean_data_tools#21)".
