### 2020-09-07 lnferris/ocean_data_tools#19 @lnferris
I have removed bathymetry_chord(). You are correct that it serves narrow purpose and I think it will get users into more trouble than it is worth. For bathymetry_search I have replaced the search for nearest grid cells with grid densification and interpolation as you have suggested. I think the points are ordered by the chosen reference variable rather than the order of the input (see below) but please let me know if I have misunderstood this comment.
```
% Order bathymetry_section by xref.
[xvar,xvar_inds] = sort(xvar);
bath_section = bath_section(xvar_inds);
lon_section = lon_section(xvar_inds);
lat_section = lat_section(xvar_inds);
time_section = time_section(xvar_inds);
```
I am closing this issue but feel free to reopen it if these fixes are not sufficient. Thank you for the helpful advice on this.
