### 2020-09-07 lnferris/ocean_data_tools#19 @lnferris
For bathymetry_search I replaced the search for nearest grid cells with interpolation followed by search for nearest grid cells as you have suggested.
