+++ ocean_data_tools/bathymetry_section.m
@@ -1,21 +1,20 @@
 
-function [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathy,xcoords,ycoords,xref,filled,maxdistance)
+function [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathy,xcoords,ycoords,xref,filled)
 % bathymetry_section adds global seafloor topography (Smith & Sandwell, 1997) to an existing section
-% plot from points nearest to specified coordinates
+% plot using specified coordinates
 % 
 %% Syntax
 % 
 %  [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathy,xcoords,ycoords,xref)
 %  [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathy,xcoords,ycoords,xref,filled)
-%  [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathy,xcoords,ycoords,xref,filled,maxdistance)
 % 
 %% Description 
 % 
 % [bath_section,lon_section,lat_section] =
 % bathymetry_section(bathy,xcoords,ycoords,xref) makes a section plot from 
 % bathy, where bathy is a struct of Smith & Sandwell Global Topography created
-% using bathymetry_extract.  Points are extracted nearest to each coordinate specified
-% by xcoords (longitude) and ycoords (latitude). The bathymetry section is
+% using bathymetry_extract. xcoords (longitude) and ycoords (latitude) are
+% densified to a 1/60-deg grid before bathymetry is interpolated. The bathymetry section is
 % plotted against xref; where xref = 'lon', 'lat', or a time vector of length(xcoords). The
 % extracted data is output bath_section, lon_section, lat_section, and time_section; 
 % output vectors are sorted by the selected reference axis (longitude,
@@ -26,11 +25,6 @@ function [bath_section,lon_section,lat_section,time_section] = bathymetry_sectio
 % bathymetry to be filled in black down to the x-axis (instead of a simple line).
 % Set filled=1 to turn on, filled=0 to turn off.
 %
-% [bath_section,lon_section,lat_section] =
-% bathymetry_section(bathy,xcoords,ycoords,xref,filled,maxdistance)
-% does not pull values where xcoords and ycoords are not within maxdistance (degrees) of a Global Topography
-% value. maxdistance=0.05 would pull no bathymetry at times/places further than
-% 0.05 diagonal degrees from an available Global Topography value.
 %
 %% Example 1
 % Add bathymetry to a temperature section plot from the list of coordinates
@@ -64,23 +58,9 @@ function [bath_section,lon_section,lat_section,time_section] = bathymetry_sectio
 % figure
 % bathymetry_section(bathy,xcoords,ycoords,xref)
 %
-%% Example 4
-% Plot bathymetry nearest to a list of coordinates, using time as the
-% x-axis and shading bathymetry. Only return bathymetry values for timesteps within 0.007 degrees
-% of a bathymetry node:
-% 
-% xref = [737009 737010 737011 737012 737013]; 
-% xcoords = [60 60.1 60.4 60.2 59.9]; 
-% ycoords = [10 20.1 15.0 16.1 13.7]; 
-% [bathy] = bathymetry_extract(bathymetry_dir,bounding_region([],xcoords,ycoords));
-% filled = 1;
-% maxdistance = 0.007;
-% figure
-% bathymetry_section(bathy,xcoords,ycoords,xref,filled,maxdistance)
-%
 %% Citation Info 
 % github.com/lnferris/ocean_data_tools
-% Jun 2020; Last revision: 07-Sep-2020
[…]
