### 2020-09-08 279fcab
Update bathymetry_section.md
