+++ docs/bathymetry_section.md
@@ -5,15 +5,12 @@
 ```Matlab
 [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathy,xcoords,ycoords,xref)
 [bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathy,xcoords,ycoords,xref,filled)
-[bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathy,xcoords,ycoords,xref,filled,maxdistance)
 ```
 #### Description
 
-``[bath_section,lon_section,lat_section] = bathymetry_section(bathy,xcoords,ycoords,xref)`` makes a section plot from ``bathy``, where ``bathy`` is a struct of Smith & Sandwell Global Topography created using ``bathymetry_extract``.  Points are extracted nearest to each coordinate specified by ``xcoords`` (longitude) and ``ycoords`` (latitude). The bathymetry section is plotted against ``xref``; where ``xref = 'lon'``, ``'lat'``, or a time vector of length(xcoords). The extracted data is output ``bath_section``, ``lon_section``, ``lat_section``, and ``time_section``; output vectors are sorted by the selected reference axis (longitude, latitude, or time).
+``[bath_section,lon_section,lat_section] = bathymetry_section(bathy,xcoords,ycoords,xref)`` makes a section plot from ``bathy``, where ``bathy`` is a struct of Smith & Sandwell Global Topography created using ``bathymetry_extract``. ``xcoords`` (longitude) and ``ycoords`` (latitude) are densified to a 1/60-deg grid before bathymetry is interpolated. The bathymetry section is plotted against ``xref``; where ``xref = 'lon'``, ``'lat'``, or a time vector of length(xcoords). The extracted data is output ``bath_section``, ``lon_section``, ``lat_section``, and ``time_section``; output vectors are sorted by the selected reference axis (longitude, latitude, or time).
  
 ``[bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathy,xcoords,ycoords,xref,filled)`` allows the bathymetry to be filled in black down to the x-axis (instead of a simple line). Set ``filled=1`` to turn on, ``filled=0`` to turn off.
-
-``[bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathy,xcoords,ycoords,xref,filled,maxdistance)`` does not pull values where ``xcoords`` and ``ycoords`` are not within ``maxdistance`` (degrees) of a Global Topography value. ``maxdistance=0.05`` would pull no bathymetry at times/places further than 0.05 diagonal degrees from an available Global Topography value.
 
 #### Example 1
 
@@ -52,19 +49,6 @@ ycoords = [10 20.1 15.0 16.1 13.7];
 [bathy] = bathymetry_extract(bathymetry_dir,bounding_region([],xcoords,ycoords));
 figure
 bathymetry_section(bathy,xcoords,ycoords,xref)
-```
-#### Example 4
-```Matlab
-% Plot bathymetry nearest to a list of coordinates, using time as the x-axis and shading bathymetry. Only return bathymetry values for timesteps within 0.007 degrees of a bathymetry node:
-
-xref = [737009 737010 737011 737012 737013]; 
-xcoords = [60 60.1 60.4 60.2 59.9]; 
-ycoords = [10 20.1 15.0 16.1 13.7]; 
-[bathy] = bathymetry_extract(bathymetry_dir,bounding_region([],xcoords,ycoords));
-filled = 1;
-maxdistance = 0.007;
-figure
-bathymetry_section(bathy,xcoords,ycoords,xref,filled,maxdistance)
 ```
 
 [Back](https://github.com/lnferris/ocean_data_tools#adding-bathymetry-to-existing-plots-1)
