### 2020-09-08 #2497 @lnferris
I have fixed and closed "Bathymetry plotting along a transect (lnferris/ocean_data_tools#19)".
