### 2020-09-08 lnferris/ocean_data_tools#33 @lnferris
Thanks for the advice about explicitly suggesting what to cite. I put a temporary "Cite as:...." statement in the README. I will replace this with the JOSS DOI and citation if/when the paper is accepted.
