### 2020-09-08 #2497 @lnferris
I have fixed and closed "DOI for each release  (lnferris/ocean_data_tools#33)".
