### 2020-09-08 e5fe8f0
expand allowable zrange
