### 2020-09-08 3a0e0d1
Update general_depth_subset.md
