### 2020-09-08 lnferris/ocean_data_tools#20 @lnferris
I have fixed this to use the same sign convention as the struct to be subsetted.
