### 2020-09-08 #2497 @lnferris
I have fixed and closed "general_depth_subset z-axis (lnferris/ocean_data_tools#20)".
