+++ ocean_data_tools/mocha_domain_plot.m
@@ -26,6 +26,10 @@ function mocha_domain_plot(month,variable,region)
 % 
 % See also mocha_build_profiles and mocha_simple_plot.
 
+% deal with inputs other than [-90 90 -180 180] e.g  [-90 90 20 200] 
+region(region>180) = region(region>180)- 360;
+region(region<-180) = region(region<-180)+360;
+
 url = 'http://tds.marine.rutgers.edu/thredds/dodsC/other/climatology/mocha/MOCHA_v3.nc';
 nc = ncgeodataset(url); % Assign a ncgeodataset handle.
 nc.variables % Print list of available variables. 
