### 2020-09-09 lnferris/ocean_data_tools#15 @lnferris
Fixed argo_build and tested other region-related functions (e.g. model_simple_plot, etc.) with the suggested intervals.
