### 2020-09-09 033eac6
command window output
