+++ ocean_data_tools/model_domain_plot.m
@@ -56,6 +56,7 @@ else
 end
 
 nc = ncgeodataset(source); % Assign a ncgeodataset handle.
+disp([ newline 'List of available variables:'])
 nc.variables            % Print list of available variables. 
 
 if ~any(strcmp(standard_vars,variable))    
@@ -68,7 +69,9 @@ if ~any(strcmp(standard_vars,variable))
 end
 
 sv = nc{variable}; % Assign ncgeovariable handle: 'water_u' 'water_v' 'water_temp' 'salinity'
+disp([ newline 'List of variable attributes:'])
 sv.attributes % Print ncgeovariable attributes.
+disp([ newline 'Available date range:'])
 datestr(sv.timeextent(),29) % Print date range of the ncgeovariable.
 svg = sv.grid_interop(:,:,:,:); % Get standardized (time,z,lat,lon) coordinates for the ncgeovariable.
 
