+++ ocean_data_tools/utilities/model_simple_plot_velocity.m
@@ -16,7 +16,9 @@ function [data,lat,lon] = model_simple_plot_velocity(model,nc,date,region,depth,
         return
     end
 
+    disp([ newline 'List of variable attributes:'])
     sv.attributes % Print ncgeovariable attributes.
+    disp([ newline 'Available date range:'])
     datestr(sv.timeextent(),29) % Print date range of the ncgeovariable.
     svg = sv.grid_interop(:,:,:,:); % Get standardized (time,z,lat,lon) coordinates for the ncgeovariable.
     
