### 2020-09-09 lnferris/ocean_data_tools#23 @lnferris
I have used ``disp`` to add explanatory messages prefacing the outputs.
