### 2020-09-09 #2497 @lnferris
I have fixed and closed "Unsuppressed command window output (lnferris/ocean_data_tools#23)".
