### 2020-09-09 7b01edb
do not create new figure when plotting
