+++ ocean_data_tools/model_simple_plot.m
@@ -150,7 +150,6 @@ data = data(:,lon_inds);
 
 % Plot
 
-figure; 
 pcolor(lon,lat,data); 
 shading flat
 title({sprintf('%s %.0fm',sv.attribute('standard_name'),svg.z(din));datestr(svg.time(tin))},'interpreter','none');
