+++ ocean_data_tools/utilities/model_domain_plot_velocity.m
@@ -64,8 +64,6 @@ end
 
 % Make Plot
 
-figure 
-
 if need2merge == 1
     
     u = reshape(permute(squeeze(double(sv.data(tin,:,lats:latn,lonw_A:lone_A))),[2,3,1]),[],1); % permute array to be lon x lat x dep
