+++ ocean_data_tools/woa_domain_plot.m
@@ -92,8 +92,6 @@ if lonw > lone
 end
 
 % Make Plot
-
-figure 
     
 if need2merge == 1
     
