+++ ocean_data_tools/woa_simple_plot.m
@@ -107,7 +107,6 @@ data = data(:,lon_inds);
 
 % Plot
 
-figure; 
 pcolor(lon,lat,data); 
 shading flat
 title({sprintf('%s %.0fm',sv.attribute('standard_name'),svg.z(din));path},'interpreter','none');
