+++ demos/demos.m
@@ -51,20 +51,25 @@ platform_id = 1900980;
 % argo_platform_map 
 
 annotate = 1; 
-argo_platform_map(argo,annotate) % annotate optional,  1=on 0=off
+figure
 bathymetry_plot(bathymetry_extract(bathymetry_dir,bounding_region(argo)),'2Dcontour')
+argo_platform_map(argo,annotate) % annotate optional,  1=on 0=off
+
 
 % argo_profiles
 
 variable = 'TEMP_ADJUSTED'; % See object for options.
 annotate = 1; 
+figure
 argo_profiles(argo,variable,annotate) % annotate optional,  1=on 0=off
 
 % argo_profiles_map
 
 annotate = 1; 
-argo_profiles_map(argo,annotate) % annotate optional,  1=on 0=off
+figure
 bathymetry_plot(bathymetry_extract(bathymetry_dir,bounding_region(argo)),'2Dcontour') % add bathymetry contours
+argo_profiles_map(argo,annotate) % annotate optional,  1=on 0=off
+
 
 
 %%                  bathymetry demonstration
@@ -84,6 +89,7 @@ bathymetry_plot(bathy,ptype)
 
 [cruise] = whp_cruise_build(ctdo_dir,uv_dir,wvke_dir,{'temperature'}); 
 xref = 'lon'; % 'lon' 'lat' or a time vector of length(xcoords)
+figure
 general_section(cruise,'temperature',xref,'pressure')
 xcoords = cruise.lon; % could alternatively use transect_select() to select coordinates
 ycoords = cruise.lat;
@@ -93,12 +99,14 @@ region = bounding_region([],xcoords,ycoords);
 bathymetry_section(bathy,xcoords,ycoords,xref,filled); % filled optional
 
 
+
 %%                  general demonstration
 
 % general_map
 [argo,matching_files] = argo_build(argo_dir,[-60.0 -50.0 150.0 160.0],'01-Nov-2015 00:00:00','01-Jan-2017 00:00:00',{'TEMP_ADJUSTED'});
 ptype = '2Dcontour'; % '2Dscatter' '2Dcontour'
 object = argo; % argo, cruise, hycom, mercator, woa, wod
+figure
 general_map(object,bathymetry_dir,ptype) % bathymetry_dir, ptype optional
 
 % general_depth_subset
@@ -113,6 +121,7 @@ zrange = [350 150]; % zrange should have the same sign convention as depth varia
 % general_region_subset
 
 object = argo; % % argo, cruise, hycom, mercator, woa, wod
+figure
 general_map(object)
 [xcoords,ycoords] = region_select(); % click desired  region on the figure
 [object_sub] = general_region_subset(object,xcoords,ycoords); 
@@ -136,12 +145,14 @@ xref = 'lat'; % 'lat' 'lon' 'stn';
 zref = 'PRES_ADJUSTED'; % See particular struct for options
 interpolate = 0; % 1=on 0=off
 contours = 0; % 1=on 0=off
+figure
 general_section(object,variable,xref,zref,interpolate,contours) % interpolate, contours optional
 
 % general_profiles
 
 object = argo; % argo, cruise, hycom, mercator, woa, wod
 variable = 'TEMP_ADJUSTED'; % see particular struct for options
+figure
 general_profiles(object,variable,zref)
 
 
@@ -153,6 +164,7 @@ month = 10; % Month (1 through 12).
 depth = 0;
 variable = 'temperature'; %  'temperature' 'salinity'
 region = [34 42  -80 -70]; % [30 48 -80 -58]
+figure
 mocha_simple_plot(month,depth,variable,region)
 
 % mocha_build_profiles
@@ -160,12 +172,16 @@ mocha_simple_plot(month,depth,variable,region)
[…]
