### 2020-09-10 4763a99
avoid setting plot options
