### 2020-09-10 lnferris/ocean_data_tools#14 @lnferris
I have fixed the plotting functions so that they do not create new figures (instead allowing the user to target an existing axis), change the stacking order, add grid lines, add a surprise margin to bounding region, nor sneakily alter the colormap.
