### 2020-09-10 #2497 @lnferris
I have fixed and closed "Plotting function customization (lnferris/ocean_data_tools#14)".
