### 2020-09-10 bb3c4d2
Update paper.md
