### 2020-09-10 8a4810f
Update paper.md
