+++ paper/paper.md
@@ -48,8 +48,8 @@ from model data. This may be used as a cruise planning tool or to facilitate
 comparison of observations (such as those from underwater glider surveys) with 
 model output. Some ``ocean_data_tools`` functions employ ``nctoolbox`` [@nctoolbox].
 
-There exist several quality ocean and/or climate related Matlab toolboxes such as Climate Data Toolbox for Matlab [@Greene:2020], those part of SEA-MAT: Matlab Tools for Oceanographic Analysis (https://sea-mat.github.io/sea-mat/), and Gibbs-SeaWater (GSW) Oceanographic Toolbox [@McDougall:2011]. However, there are no other documented and designed-to-be-shared toolboxes filling the same data exploration niche as this one. ``ocean_data_tools`` is unique in encouraging the user to invoke a variety of freely-available data into their exploration and does not expect the user to provide privately-collected measurements or privately-generated model output. ``ocean_data_tools`` has already been used for data exploration in support of scientific publications [@Bemis:2020] 
-and [@Crear:2020]. This toolbox is built for extensibility; the plan 
+There are several high-quality ocean and/or climate related Matlab toolboxes such as Climate Data Toolbox for Matlab [@Greene:2020], those part of SEA-MAT: Matlab Tools for Oceanographic Analysis (https://sea-mat.github.io/sea-mat/), and Gibbs-SeaWater (GSW) Oceanographic Toolbox [@McDougall:2011]. However, there are no other documented and designed-to-be-shared toolboxes filling the same data exploration niche as this one. ``ocean_data_tools`` is unique in encouraging the user to invoke a variety of freely-available data into their exploration and does not expect the user to provide privately-collected measurements or privately-generated model output. ``ocean_data_tools`` has already been used for data exploration in support of scientific publications [@Bemis:2020] 
+and [@Crear:2020]. This toolbox is built for extensibility; the objective is 
 is to welcome contributors and continuously add support for additional datasets such as Remote Sensing 
 Systems (http://www.remss.com/) products and European Centre for Medium-Range 
 Weather Forecasts (ECMWF) products. The source code for ``ocean_data_tools`` has
