+++ paper/paper.bib
@@ -59,3 +59,13 @@
     volume = 29,
     doi = {10.5670/oceanog.2016.66},
 }
+
+@article{Greene:2019,
+  	Author = {{Greene}, C. ~A. and {Thirumalai}, K. and {Kearney}, K. ~A. and {Delgado}, J. ~M. and {Schwanghart}, W. and {Wolfenbarger}, N. ~S. and {Thyng}, K. ~M. and {Gwyther}, D. ~E. and {Gardner}, A. ~S. and {Blankenship}, D. ~D.},
+  	Journal = {Geochemistry, Geophysics, Geosystems},
+  	Title = {{The Climate Data Toolbox for MATLAB}},
+  	Year = 2019,
+    doi = {10.1029/2019GC008392},
+}
+
+
