### 2020-09-10 6bdbd3f
Update paper.md
