### 2020-09-10 cfb668f
Update paper.md
