### 2020-09-10 86716a6
Update paper.md
