### 2020-09-10 3baabe5
Update paper.md
