### 2020-09-10 3e47dfd
Update paper.md
