### 2020-09-10 1c06cdd
Update paper.md
