+++ paper/paper.md
@@ -28,14 +28,6 @@ solves this problem by allowing the user to transform common oceanographic
 data sources into uniform structure arrays, call general functions on these structure arrays, 
 perform custom calculations, and make graphics. 
 
-![Building a virtual cruise from the Operational Mercator global ocean
-analysis and forecast system at 1/12 degree with 3D bathymetry [@Smith:1997]. 
-Showing (a) a 3D velocity plot created using ``model_domain_plot``, (b) 
-virtual cruise selection using ``transect_select``, and ``model_build_profiles``, 
-(c) coordinates of the resulting uniform structure array, and (d) a temperature section 
-plotted using ``general_section`` with ``bathymetry_section``. Three of the 
-subplots use colormaps from cmocean [@Thyng:2016]. \label{fig:1}](figure.png)
-
 # Summary
 
 Structure arrays, the common currency of ``ocean_data_tools``, are more user-friendly than the native data storage underlying many of the datasets because they allow the user to neatly group related data of any type or size into containers called fields. Both the structure array and its fields are mutable, and data is directly visible and accessible in the Matlab workspace (unlike NetCDF which requires a function call to read variables).
@@ -53,6 +45,14 @@ draws transects on a map (or passes coordinates as an argument) to build vertica
 from model data. This may be used as a cruise planning tool or to facilitate 
 comparison of observations (such as those from underwater glider surveys) with 
 model output. Some ``ocean_data_tools`` functions employ ``nctoolbox`` [@nctoolbox].
+
+![Building a virtual cruise from the Operational Mercator global ocean
+analysis and forecast system at 1/12 degree with 3D bathymetry [@Smith:1997]. 
+Showing (a) a 3D velocity plot created using ``model_domain_plot``, (b) 
+virtual cruise selection using ``transect_select``, and ``model_build_profiles``, 
+(c) coordinates of the resulting uniform structure array, and (d) a temperature section 
+plotted using ``general_section`` with ``bathymetry_section``. Three of the 
+subplots use colormaps from cmocean [@Thyng:2016]. \label{fig:1}](figure.png)
 
 There are several high-quality ocean and/or climate related Matlab toolboxes such as Climate Data Toolbox for Matlab [@Greene:2019], those part of SEA-MAT: Matlab Tools for Oceanographic Analysis (https://sea-mat.github.io/sea-mat/), and Gibbs-SeaWater (GSW) Oceanographic Toolbox [@McDougall:2011]. However, there are no other documented and designed-to-be-shared toolboxes filling the same data exploration niche as this one. ``ocean_data_tools`` is unique in encouraging the user to invoke a variety of freely-available data into their exploration and does not expect the user to provide privately-collected measurements or privately-generated model output. It connects users to specific, well-documented data sources. ``ocean_data_tools`` has already been used for data exploration in support of scientific publications [@Bemis:2020] 
 and [@Crear:2020]. This toolbox is built for extensibility; the objective is 
