+++ paper/paper.md
@@ -38,7 +38,8 @@ The workflow of ``ocean_data_tools`` is to build uniform structure arrays (e.g. 
 ``cruise``, ``hycom``, ``mercator``, ``woa``, ``wod``) from raw datasets and 
 call general functions on these structure arrays to map, subset, or plot. Functions with 
 the ``\_build`` suffix load raw data into uniform structure arrays. Structure arrays are 
-compatible with all ``general_`` functions, as well as functions in the commonly-used
+compatible with all ``general_`` functions, and serve to neatly contain the data for use with 
+custom user-defined calculations or other toolboxes such as the commonly-used
 Gibbs-SeaWater (GSW) Oceanographic Toolbox [@McDougall:2011]. One application of the ``\_build`` 
 feature is to create virtual cruises from model output \autoref{fig:1}. The user
 draws transects on a map (or passes coordinates as an argument) to build vertical profiles 
