### 2020-09-10 lnferris/ocean_data_tools#27 @lnferris
I have removed the compatibility claims! Reference to GSW is now on the same playing field as the suggestion to apply custom calculations after data has been neatly packed into structs.
