### 2020-09-10 5c345e4
Update paper.md
