### 2020-09-10 lnferris/ocean_data_tools#26 @lnferris
I have expanded the installation instructions, added information about which datasets are manual download vs. remote access, and fixed broken links.
