### 2020-09-10 #2497 @lnferris
I have fixed and closed "Installation instructions (lnferris/ocean_data_tools#26)".
