### 2020-09-10 lnferris/ocean_data_tools#34 @lnferris
I have addressed #26 and am currently addressing #29. #28 is probably the only of your many helpful suggestions that I am going to push back on. The reason is that this is intended to be a living project, and I think it will be complicated to have to change the documentation in 3 places (embedded in functions, markdown, html) every time contributor someone adds or tweaks a function. I want contributing to be as accessible as possible and don't want to add a redundant place where documentation has to be changed/expanded every time this happens.

Chad Greene (who was initially reviewing this) opened an issue requiring more documentation and suggested it could take the form of either a documentation page accessible in the Matlab Documentation Viewer or it could just detailed in the README.md. I addressed this by implementing linked Markdown files within the GitHub ecosystem, which I prefer to the Matlab ecosystem because it can more easily be expanded to documentation of shell scripts, python, etc.
