+++ docs/argo_build.md
@@ -7,8 +7,17 @@
 ```
 #### Description
 
-``[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)`` searches pathway ``argo_dir`` for profiles meeting the search criteria ``region``, ``start_date``, and ``end_date``. Profiles are loaded into the struct array ``argo`` with all variables specified in ``variable_list``. Variables PLATFORM_NUMBER, LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically. Files containing matching profiles are listed in ``matching_files``.
+``[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)`` searches pathway ``argo_dir`` for profiles meeting the search criteria ``region``, ``start_date``, and ``end_date``. Profiles are loaded into the struct array ``argo`` with all variables specified in ``variable_list``. Files containing matching profiles are listed in ``matching_files``.
 
+``argo_dir`` is a character array search path with wildcards. The search path should be the path to the netcdf files themselves, not their directory. 
+
+``region`` is a vector containing the bounds [S N W E] of the search region, with limits [-90 90 -180 180]. Limits may cross the dateline e.g. [35 45 170 -130].
+
+``start_date`` and ``end_date`` are date strings in format ``'dd-mmm-yyyy HH:MM:SS'``.
+
+``argo`` is a uniform struct containing data from the profiles matching the region and date criteria. Some data is included automatically while some must be specificed. The variables PLATFORM_NUMBER, LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically. Additional variables must be specified in ``variable_list``, a cell array where each element is the string name of a variable.
+
+``matching_files`` is a string array where each string is the full path to a file which contained a profile matching the region and date criteria.
 
 #### Example 1
 
