+++ ocean_data_tools/argo_profiles.m
@@ -17,6 +17,9 @@ function argo_profiles(argo,variable,annotate)
 % markers by default annotate=0. set annotate=1 to turn on annotation. The 
 % annotations of profiles correspond to those of argo_profiles_map called
 % on the same struct.
+%
+% variable is the string name of the field (of argo) to be plotted as a vertical
+% profile
 % 
 %% Example 1
 % Plot temperature vertical profiles in argo:
