### 2020-09-11 3e39b87
Update bathymetry_extract.md
