### 2020-09-11 806f2f5
Update bounding_region.md
