### 2020-09-11 04055e8
Update bounding_region.md
