+++ docs/bathymetry_section.md
@@ -12,6 +12,10 @@
  
 ``[bath_section,lon_section,lat_section,time_section] = bathymetry_section(bathy,xcoords,ycoords,xref,filled)`` allows the bathymetry to be filled in black down to the x-axis (instead of a simple line). Set ``filled=1`` to turn on, ``filled=0`` to turn off.
 
+``xcoords`` and ``ycoords`` are vectors of coordinates. Rows or columns are fine, and both -180/180 or 0/360 notation are fine.
+
+When ``xref`` is a time vector, it must be of ``length(xcoords)`` and elements of the vector must be datenums. Otherwise just set ``xref = 'lon'`` or  ``xref = 'lat'``. 
+
 #### Example 1
 
 ```Matlab
