### 2020-09-11 c88c5e7
Update bathymetry_extract.md
