### 2020-09-11 13af04b
Update general_remove_duplicates.md
