### 2020-09-11 a0f8732
Update general_depth_subset.md
