### 2020-09-11 414dd9c
Update general_map.md
