+++ docs/general_profiles.md
@@ -9,6 +9,10 @@ general_profiles(object,variable,zref)
 
 ``general_profiles(object,variable,zref)`` plots vertical profiles of the specified ``variable`` in struct ``object`` as a function of the depth field specified by ``zref``; where ``object`` is a struct created by any of the ``_build`` functions in ocean_data_tools (e.g. ``argo``, ``cruise``, ``hycom``, ``mercator``, ``woa``, ``wod``) and ``variable`` is a field name.
 
+``variable`` is the string name of the field (of ``object``) to be plotted as the x-variable of the vertical profile
+
+``zref`` is the string name of the field (of ``object``) to be plotted as the depth variable of the vertical profile
+
 #### Example 1
 
 
