### 2020-09-11 63dbe86
Update general_section.md
