+++ docs/mocha_build_profiles.md
@@ -8,10 +8,13 @@
 ```
 #### Description
 
-``[mocha] = mocha_build_profiles(month,xcoords,ycoords)`` builds a struct of profiles from the MOCHA Mid-Atlantic Bight climatology, pulling profiles nearest to coordinates specified by ``xcoords`` and ``ycoords``. The calendar month is specified by ``month``.
+``[mocha] = mocha_build_profiles(month,xcoords,ycoords)`` builds a unform struct, ``mocha`` of profiles from the MOCHA Mid-Atlantic Bight climatology, pulling profiles nearest to coordinates specified by ``xcoords`` and ``ycoords``. The calendar month is specified by ``month``.
 
 ``[mocha] = mocha_build_profiles(month,xcoords,ycoords,zgrid)`` depth-interpolates the profiles to a vertical grid of ``zgrid``, in meters. ``zgrid=2`` would produce profiles interpolated to 2 meter vertical grid.
 
+``xcoords`` and ``ycoords`` are vectors of coordinates. Rows or columns are fine, and both -180/180 or 0/360 notation are fine.
+
+``month`` is an integer between 1 (January) and 12 (December).
 
 #### Example 1
 
