### 2020-09-11 b5c3de3
Update mocha_simple_plot.md
