### 2020-09-11 d00e3b4
Update mocha_domain_plot.md
