### 2020-09-11 08a0d4c
Update model_simple_plot.md
