### 2020-09-11 957164b
Update model_simple_plot.md
