### 2020-09-11 a9461a6
Update model_domain_plot.md
