+++ ocean_data_tools/general_profiles.m
@@ -13,6 +13,12 @@ function general_profiles(object,variable,zref)
 % specified by zref; where object is a struct created by any of 
 % the _build functions in ocean_data_tools (e.g. argo, cruise, hycom, mercator,
 % woa, wod) and variable is a field name.
+%
+% variable is the string name of the field (of object) to be plotted as the 
+% x-variable of the vertical profile
+%
+% zref is the string name of the field (of object) to be plotted as the 
+% depth variable of the vertical profile
 % 
 %% Example 1
 % Plot temperature vertical profiles in argo:
