+++ ocean_data_tools/general_section.m
@@ -22,6 +22,15 @@ function  general_section(object,variable,xref,zref,interpolate,contours)
 % general_section(object,variable,xref,zref,interpolate,contours) adds
 % contours to the section plot. contours=1 for on, contours=0 for off.
 %
+% variable is the string name of the field (of object) to be plotted as the 
+% color variable of the section plot
+%
+% zref is the string name of the field (of object) to be plotted as the 
+% depth variable of the section plot
+%
+% xref is the string name of the field (of object) to be plotted as the 
+% horizontal variable of the section plot
+%
 %% Example 1
 % Plot a temperature section from a hycom struct:
 % 
