### 2020-09-11 a54f48b
Update region_select.md
