+++ docs/region_select.md
@@ -9,6 +9,8 @@
 
 ``[xcoords,ycoords] = region_select()`` creates a list of x and y coordinates (which represent vertices of a polygon) selected by clicking stations on an existing (latitude vs. longitude) plot, returning them as ``xcoords`` and ``ycoords``.
 
+``xcoords`` and ``ycoords`` are vectors of coordinates representing vertices of a polygon. -180/180 or 0/360 notation will match that of the existing plot.
+
 #### Example 1
 
 ```Matlab
