### 2020-09-11 0c353bb
Update whp_cruise_build.md
