### 2020-09-11 36d8e29
Update whp_cruise_build.md
