+++ docs/whp_cruise_build.md
@@ -20,8 +20,11 @@ The paths used as arguments should point to data from the *same oceanographic cr
 ``wvke_dir`` is a character array path to all files in the directory. 
 
 Example paths: 
+
 ctdo_dir = '/Users/lnferris/Documents/S14/ctd/\*.nc'; 
+
 uv_dir = '/Users/lnferris/Documents/S14/whp_cruise/uv/\*.nc';
+
 wvke_dir = '/Users/lnferris/Documents/S14/whp_cruise/wvke/';
 
 #### Example 1
