### 2020-09-11 c77c124
Update whp_cruise_build.md
