### 2020-09-12 3114a7d
Update woa_simple_plot.md
