+++ ocean_data_tools/transect_select.m
@@ -16,6 +16,12 @@ function [xcoords,ycoords] = transect_select(densify)
 % stations with the multiplier densify. densify=10 would fill in 10
 % stations for every station clicked using linear interpolation of complex
 % coordinates
+%
+% densify should be an integer. If it is not an integer it will be rounded
+% to an integer.
+%
+% xcoords and ycoords are vectors of coordinates representing a polygonal chain.
+% -180/180 or 0/360 notation will match that of the existing plot.
 % 
 %% Example 1
 % Generate a list of coordinates by clicking a HYCOM velocity plot: 
@@ -51,7 +57,7 @@ if nargin >= 1
 % interpolate complex coordinates
 xy_coords = xcoords + ycoords*1j; % z = x+iy
 v = xy_coords;
-xq  = linspace(1,length(xy_coords),length(xy_coords)*densify);
+xq  = linspace(1,length(xy_coords),length(xy_coords)*round(densify));
 xxyy_coords = interp1(v,xq);
 xcoords = real(xxyy_coords); 
 ycoords = imag(xxyy_coords);
