### 2020-09-12 lnferris/ocean_data_tools#29 @lnferris
I have gone through every file and expanded the in-file documentation (as well as the browser based documentation linked from the main readme).
