### 2020-09-12 e19e655
input parsing and validation
