### 2020-09-12 32f0c59
input parsing and validation
