### 2020-09-12 lnferris/ocean_data_tools#10 @lnferris
I have added input validation to functions and modified argo_build as suggested. Thanks for the advice!
