### 2020-09-12 #2497 @lnferris
I have fixed and closed "Input parsing and validation (lnferris/ocean_data_tools#10)".
