### 2020-09-12 #2497 @lnferris
Hello @kthyng @kakearney @castelao! I believe I have addressed the required issues and some of the suggested issues (hope to get to the others when time allows). The only one marked as required which I have chosen not to implement is #28; there is an explanation about this in the issue itself. Please let me know if there are more changes and or/tasks needed from my end. Thanks again for your detailed reviews.
