### 2020-09-12 0a374d1
Update whp_cruise_build.md
