### 2020-09-19 5ccf8a9
Update paper.md
