### 2020-09-19 67e7ad8
Update paper.md
