### 2020-09-19 7ccb3b0
Update paper.md
