+++ paper/paper.md
@@ -45,7 +45,7 @@ feature is to create virtual cruises from model output \autoref{fig:1}. The user
 draws transects on a map (or passes coordinates as an argument) to build vertical profiles 
 from model data. This may be used as a cruise planning tool, to facilitate 
 comparison of observations with model output, or to support decision-making in underwater glider 
-piloting (e.g. use model forecasts to inform ballasting or adjust flight for ocean currents).  Some ``ocean_data_tools`` functions
+piloting (using model forecasts to inform ballasting or adjust flight for ocean currents).  Some ``ocean_data_tools`` functions
 employ ``nctoolbox`` [@nctoolbox].
 
 ![Building a virtual cruise from the Operational Mercator global ocean
