+++ demos/pathdef.m
@@ -1,0 +1,1773 @@
+function p = pathdef
+%PATHDEF Search path defaults.
+%   PATHDEF returns a string that can be used as input to MATLABPATH
+%   in order to set the path.
+
+  
+%   Copyright 1984-2016 The MathWorks, Inc.
+
+
+% DO NOT MODIFY THIS FILE.  IT IS AN AUTOGENERATED FILE.  
+% EDITING MAY CAUSE THE FILE TO BECOME UNREADABLE TO 
+% THE PATHTOOL AND THE INSTALLER.
+
+p = [...
+%%% BEGIN ENTRIES %%%
+     '/Users/lnferris/Desktop/nctoolbox-master:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/cdm:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/cdm/utilities:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/cdm/utilities/graphics:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/cdm/utilities/interp:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/cdm/utilities/misc:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/cdm/utilities/njcompatability:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/cdm/utilities/search:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/cdm/utilities/slicing:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/cdm/utilities/units:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/cdm/utilities/vis:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/demos:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/demos/contrib:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/demos/html:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/docs:', ...
+     '/Users/lnferris/Desktop/nctoolbox-master/java:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/branches:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/hooks:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/info:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/logs:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/logs/refs:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/logs/refs/heads:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/logs/refs/remotes:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/logs/refs/remotes/origin:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/objects:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/objects/00:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/objects/01:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/objects/02:', ...
+     '/Users/lnferris/Documents/GitHub/ocean_data_tools/.git/objects/03:', ...
[…]
