### 2020-09-22 0271a55
Update general_section.md
