### 2020-09-22 063f5a5
Update general_section.md
