### 2020-09-22 aca1734
Delete pathdef.m
