### 2020-09-22 lnferris/ocean_data_tools#24 @lnferris
Thanks for the suggestion! I have added this feature to bathymetry_section and general_section.
