+++ ocean_data_tools/general_section.m
@@ -31,7 +31,8 @@ function  general_section(object,variable,xref,zref,interpolate,contours)
 % xref is the string name of the field (of object) to be plotted as the 
 % horizontal variable of the section plot, usually 'stn', 'lat', or 'lon'.
 % Alteratively pass xref = 'km' to plot in along-track distance, assuming 
-% spherical earth.
+% spherical earth. If field object.km already exists, this will be used 
+% instead of calculating along-track distance.
 %
 %% Example 1
 % Plot a temperature section from a hycom struct:
