### 2020-09-22 #2497 @lnferris
I have fixed and closed "Plot versus distance along transect (lnferris/ocean_data_tools#24)".
