### 2020-09-23 e0db886
Update general_section.md
