+++ docs/model_build_profiles.md
@@ -57,7 +57,7 @@ model_simple_plot(model_type,source,date,variable,region,depth,arrows)
 
 % Click stations on the plot to create a coordinate list:
 
-[xcoords,ycoords] = transect_select(10); % click desired transect on the figure, densify selection by 10x 
+[xcoords,ycoords] = transect_select('densify',10); % click desired transect on the figure, densify selection by 10x 
 
 ```
 <img src="https://user-images.githubusercontent.com/24570061/88411026-3c1c3180-cda5-11ea-81d3-d3b656315464.png" width="600">
