### 2020-09-23 5c3f5c6
Update transect_select.md
