### 2020-09-23 e228aa3
Update transect_select.md
