### 2020-09-23 393a055
Update transect_select.md
