### 2020-09-23 fb5d5ce
Update transect_select.md
