### 2020-09-23 fd5ac3c
Update transect_select.md
