+++ ocean_data_tools/transect_select.m
@@ -1,24 +1,33 @@
-function [xcoords,ycoords] = transect_select(densify)
+function [xcoords,ycoords] = transect_select(scheme,value)
 % transect_select creates a list of x and y coordinates (which represent a 
 % transect) selected by clicking stations on a plot
 % 
 %% Syntax
 % 
 %  [xcoords,ycoords] = transect_select() 
-%  [xcoords,ycoords] = transect_select(densify)
+%  [xcoords,ycoords] = transect_select(scheme,value)
 % 
 %% Description 
 % 
 % [xcoords,ycoords] = transect_select() creates a list of x and y coordinates selected by
 % clicking stations on an existing (latitude vs. longitude) plot, returning them as xcoords and ycoords
 %
-% [xcoords,ycoords] = transect_select(densify) auto-generates additional
-% stations with the multiplier densify. densify=10 would fill in 10
+% [xcoords,ycoords] = transect_select(scheme,value) auto-generates additional
+% stations with based on the scheme chosen. scheme = 'densify' auto-generates 
+% additional stations with the multiplier value; value=10 would fill in 10
 % stations for every station clicked using linear interpolation of complex
-% coordinates
+% coordinates. scheme = 'spacing' auto-generates additional stations with
+% the specified spacing value, where value is the longitude or latitude 
+% spacing in degrees; value=0.5 would fill in stations such that stations 
+% are 0.5 degrees apart. 
 %
-% densify should be an integer. If it is not an integer it will be rounded
-% to an integer.
+% If scheme = 'densify', value (no units) should be an integer. If it is not an integer 
+% it will be rounded to an integer.
+%
+% If scheme = 'spacing', value (in degrees) should be single, double, or integer and 
+% represents the prescribed grid spacing between auto-generated stations in 
+% degrees. The spacing criterion is applied to each dimension separately and 
+% is not the diagonal displacement.
 %
 % xcoords and ycoords are vectors of coordinates representing a polygonal chain.
 % -180/180 or 0/360 notation will match that of the existing plot.
@@ -33,35 +42,62 @@ function [xcoords,ycoords] = transect_select(densify)
 % region = [-5.0, 45.0 ,160,-150 ];      
 % depth = -150;                                                   
 % model_simple_plot(model_type,source,date,variable,region,depth);
-% [xcoords,ycoords] = transect_select(10); % click desired transect on the figure, densify selection by 10x 
+% [xcoords,ycoords] = transect_select('spacing',0.5); % click desired transect on the figure, densify selection to 0.5 degree spacing
 %
 %% Citation Info 
 % github.com/lnferris/ocean_data_tools
-% Jun 2020; Last revision: 20-Jul-2020
+% Jun 2020; Last revision: 22-Sep-2020
 % 
 % See also region_select.
 
 
 if nargin >= 1
     
-assert(isnumeric(densify) & length(densify)==1,'Error: densify must be an integer.');
-
-waitfor(msgbox(['Click points to draw a transect. Double-click to end the transect. The selected points will be interpolated such that ',num2str(densify),' interspersed stations will be created for every 1 point selected.']));
+    if nargin ==1
+                disp('Error: Missing second argument.')
+        return
+    end
+        
+    if strcmp(scheme,'densify')
+        
+        assert(isnumeric(value) & length(value)==1,'Error: value must be an integer.');
[…]
