### 2020-09-23 lnferris/ocean_data_tools#25 @lnferris
Users can now specify density in terms of distance.
