### 2020-09-23 #2497 @lnferris
I have fixed and closed "transect_select point density (lnferris/ocean_data_tools#25)".
