### 2020-09-23 c20626b
build function for gliders.ioos.us/erddap
