### 2020-09-23 5731b02
build function for gliders.ioos.us/erddap
