### 2020-09-23 39bb612
Update glider_build.md
