+++ docs/glider_build.md
@@ -1,29 +1,40 @@
-### argo_build
+### glider_build
 
 #### Syntax
 
 ```Matlab
-[argo,matching_files] = argo_build(argo_dir)
-[argo,matching_files] = argo_build(argo_dir,region)
-[argo,matching_files] = argo_build(argo_dir,region,start_date)
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date)
-[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)
+[glider] = glider_build(glider_dir)
+[glider] = glider_build(glider_dir,variable_list)
 ```
 #### Description
 
-``[argo,matching_files] = argo_build(argo_dir,region,start_date,end_date,variable_list)`` searches pathway ``argo_dir`` for profiles meeting the search criteria ``region``, ``start_date``, and ``end_date``. Profiles are loaded into the struct array ``argo`` with all variables specified in ``variable_list``. Files containing matching profiles are listed in ``matching_files``.
+``[glider] = glider_build(glider_dir,variable_list)`` loads data from ``glider_dir`` into struct array ``glider``. Glider profiles are loaded into the struct with all variables specified in ``variable_list``. 
 
-The only required argument is ``argo_dir``. The default state is to load all profiles in path ``argo_dir``, writing variables TEMP_ADJUSTED and PSAL_ADJUSTED into the uniform struct ``argo``.
+The only required argument is ``glider_dir``. The default state is to load the following science variables:
+```
+'profile_id'
+'time'
+'latitude'
+'longitude'
+'precise_time'
+'depth'
+'pressure'
+'temperature'
+'conductivity'
+'salinity'
+'density'
+'precise_lat'
+'precise_lon'
+'time_uv'
+'lat_uv',
+'lon_uv'
+'u'
+'v' 
+```
 
-``argo_dir`` is a character array search path with wildcards. The search path should be the path to the netcdf files themselves, not their directory. 
+This default list can be overridden by passing a user-defined ``variable_list``, a cell array where each element is the string name of a variable. 
 
-``region`` is a vector containing the bounds [S N W E] of the search region, with limits [-90 90 -180 180]. Limits may cross the dateline e.g. [35 45 170 -130].
-
-``start_date`` and ``end_date`` are date strings in format ``'dd-mmm-yyyy HH:MM:SS'``.
-
-``argo`` is a uniform struct containing data from the profiles matching the region and date criteria. Some data is included automatically while some must be specificed. The variables PLATFORM_NUMBER, LONGITUDE, LATITUDE, JULD, and PRES_ADJUSTED are included automatically. Additional variables must be specified in ``variable_list``, a cell array where each element is the string name of a variable.
-
-``matching_files`` is a string array where each string is the full path to a file which contained a profile matching the region and date criteria.
+``glider_dir`` is a character array search path to a single netcdf file downloaded from gliders.ioos.us/erddap). The path should be the path to the netcdf file itself, not its directory. 
 
 #### Example 1
 
@@ -32,26 +43,24 @@ The only required argument is ``argo_dir``. The default state is to load all pro
 
 % Get variable information:
 
-argo_dir = '/Users/lnferris/Documents/GitHub/ocean_data_tools/data/argo/*profiles*.nc';
-listing = dir(argo_dir); 
-ncdisp([listing(1).folder '/' listing(1).name]) % Peek at netCDF header info to inform choice of variable_list.
[…]
