### 2020-09-23 e7d8d9d
expand glider example
