+++ README.md
@@ -10,10 +10,6 @@ ocean_data_tools simplifies the process of extracting, formatting, and visualizi
 
 Find a bug, have a question, or want to contribute? Open an issue or email lnferris@alum.mit.edu.
 
-*Cite as:*
-
-*Ferris, L., (2020): ocean_data_tools: A MATLAB toolbox for interacting with bulk freely-available oceanographic data. July. Zenodo. doi:10.5281/zenodo.3353610.*
-
 ### [Getting Started](#getting-started-1)
 
 ### [Dependencies](#dependencies-1)
@@ -21,6 +17,8 @@ Find a bug, have a question, or want to contribute? Open an issue or email lnfer
 ### [Contents](#contents-1)
 
 ### [Finding Data](#finding-data-1)
+
+### [Citing ODT](#citing-odt-1)
 
 ## Getting Started
 
@@ -194,4 +192,10 @@ Functions build the World Ocean Atlas url at maximum resolution based on argumen
 #### wod<sup>1</sup>
 
 Search the [World Ocean Database](https://www.nodc.noaa.gov/OC5/SELECT/dbsearch/dbsearch.html) and select products.
+
+## Citing ODT
+
+Please cite as:
+
+*Ferris, L., (2020): ocean_data_tools: A MATLAB toolbox for interacting with bulk freely-available oceanographic data. July. Zenodo. doi:10.5281/zenodo.3353610.*
 
