### 2020-09-24 lnferris/ocean_data_tools#34 @kthyng
> it will be complicated to have to change the documentation in 3 places (embedded in functions, markdown, html) every time contributor someone adds or tweaks a function. I want contributing to be as accessible as possible and don't want to add a redundant place where documentation has to be changed/expanded every time this happens.

I am not familiar with the Matlab documentation discussed in #28, but generally speaking, all functions should have their own documentation which changes if that particular function changes.

I am guessing that the markdown and html subsequently mentioned is for more of a narrative documentation. Presumably there is a way to create the html pages from the markdown pages, so there wouldn't be a need to update both independently.

We should have @kakearney here too since this is related to another issue.
