### 2020-09-24 #2497 @kthyng
Ok I see issues #28 and #34 which I have commented on there. I would like to hear from the reviewers there if possible @castelao @kakearney.

@kakearney is #30 meant to be a suggestion?

Generally I'd like to hear from both reviewers about their responses to the author responses. Thank you.
