### 2020-09-24 #2497 @lnferris
@kthyng, thanks for the update; sounds good! Also, I added a new function (glider_build, for US IOOS archived glider data) since they reviewed. I documented it the same way as the others, but let me know if anyone has an issue with this.
