### 2020-09-24 #2497 @kakearney
Regarding the highlighted issues, lnferris/ocean_data_tools#28 was only a suggestion on my part.    lnferris/ocean_data_tools#30 was more of a requirement; it looks like the recent updates do add much better input parsing and validation, but I have not yet had a chance to run any tests of my own (and likely won't for another two weeks due to workload).

More details regarding 28:

Within the Matlab desktop application, function help can be provided in two forms, typically called help and doc files.  The "help" function displays plain-text info, and pulls that info from the header of each function/script file; I consider this (with description, syntax, input/output parameter descriptions, etc.) to be mandatory for shared code, and a brief perusal suggests that the updates associated with lnferris/ocean_data_tools#29 now fulfill this requirement.

More in-depth documentation can be accessed via Matlab's "doc" function, which loads html-formatted files that typically repeat the in-file documentation and also provide further context, examples, etc; all of Matlab's native functions include these but they aren't very common in third-party toolboxes.  So I wouldn't say it's a requirement to add it to this toolbox.  However, Ferris has already done the work creating markdown documentation for all of the functions, and I would encourage them to reconsider converting it to html as time permits in the future.  End users are far more likely to browse documentation from within the Matlab help viewer (i.e. doc) than they are to revisit the GitHub repo.
