### 2020-09-24 #2497 @kakearney
Most of my own documentation is written in Matlab and converted directly to
html via Matlab's publish function.  I also use my own publishreadme
<https://github.com/kakearney/publishreadme-pkg> Matlab function to
publish marked-up
Matlab scripts
<https://www.mathworks.com/help/matlab/matlab_prog/marking-up-matlab-comments-for-publishing.html>
directly
to both html and Github-flavored markdown.  I don't have specific
recommendations for going from markdown to html.

On Thu, Sep 24, 2020 at 12:23 PM Laur Ferris <notifications@github.com>
wrote:

> @kakearney <https://github.com/kakearney> Do you recommend a particular
> markdown-to-html converter?
>
