+++ docs/general_profiles.md
@@ -1,4 +1,4 @@
-### argo_build
+### general_profiles
 
 #### Syntax
 
