### 2020-09-25 lnferris/ocean_data_tools#28 @lnferris
I have added Matlab documentation.
