### 2020-09-25 lnferris/ocean_data_tools#34 @lnferris
I have implemented documentation accessible in the Matlab Documentation Viewer https://github.com/lnferris/ocean_data_tools/issues/28.
