### 2020-09-25 #2497 @lnferris
@kakearney I implemented the Matlab documentation and mentioned the availability of this documentation under "Accessing Help" in the GitHub readme. For future reference, I used [grip](https://github.com/joeyespo/grip) to convert markdown to html.
