### 2020-09-30 lnferris/ocean_data_tools#36 @castelao
DOIs of the datasets
This issue is part of the JOSS review ([openjournals/joss-reviews#2497](https://github.com/openjournals/joss-reviews/issues/2497))

Requirement: The paper and the documentation should include, when available, the DOI of the datasets used, such as Argo, WOA, WOD, ... I'm sorry that I missed this point before.

Suggestion: Include that in the DOI record created by Zenodo too. It could be done manually on every release or automatically by defining a [.zenodo.json](https://github.com/castelao/inception/blob/master/.zenodo.json) file.
