### 2020-09-30 lnferris/ocean_data_tools#37 @castelao
Community guidelines
This issue is part of the JOSS review ([openjournals/joss-reviews#2497](https://github.com/openjournals/joss-reviews/issues/2497))

One of the requirements for JOSS is: "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

If that already exists, could you point it to me, please? And maybe make it more easily visible? That would help people interested in helping you with your package. You don't need to teach the basics, it's OK to asume that this person would master Matlab and git, but tell us how is the way that you would like things to be done, with more details than simply "open an issue". Which kind of functionalities would you consider to merge in your package? If I create a new function, what would be the basic requirements for you to consider merging that? This can save you time in the long run.
