+++ README.md
@@ -53,9 +53,9 @@ To access help, run the command ``doc ocean_data_tools``.
 
 ## How to Contribute
 
-* Find a bug in the code? Open an 'issue' to notify contributors and create an official record.
-* Want to make changes or add a new function? First 'fork' the repository (make your own separate copy), then make changes, then open a 'pull request'. Once approved, it can be merged into the master branch. If you wish to inquire beforehand whether your contribution fits the scope of ocean_data_tools, open an issue or email lnferris@alum.mit.edu. 
+* Want to make changes or add a new function? First 'fork' the repository (make your own separate copy), then make changes, then open a 'pull request'. Once approved, it can be merged into the master branch. If you wish to chat beforehand whether your contribution fits the scope of ocean_data_tools, open an issue or email lnferris@alum.mit.edu. 
 * Don't use git often and don't want to remember all the terminal commands? Download GitHub for desktop <https://desktop.github.com/>.
+* Find a bug in the code? Open an 'issue' to notify contributors and create an official record.
 
 ## Contents
 
