### 2020-10-01 lnferris/ocean_data_tools#37 @lnferris
I added a new section in README "How to Contribute"
