### 2020-10-01 #2497 @lnferris
I have fixed and closed "Community guidelines (lnferris/ocean_data_tools#37)".
