+++ paper/paper.md
@@ -69,9 +69,9 @@ been archived to Zenodo with the linked DOI: [@Ferris:2020].
 | Argo floats | doi:10.17882/42182 |
 | Smith & Sandwell bathymetry | doi:10.1126/science.277.5334.1956 |
 | IOOS Glider DAC | https://gliders.ioos.us/ |
-| Mid-Atlantic Ocean Climatological Hydrographic Analysis (MOCHA) Climatology | doi:10.7282/T3XW4N4M |
+| MOCHA Climatology | doi:10.7282/T3XW4N4M |
 | HYbrid Coordinate Ocean Model (HYCOM) | https://hycom.org |
-| CMEMS Global Ocean 1/12° Physics Analysis and Forecast Updated Daily | GLOBAL_ANALYSIS_FORECAST_PHY_001_024 |
+| CMEMS Global Ocean 1/12° Physics Analysis and Forecast | GLOBAL_ANALYSIS_FORECAST_PHY_001_024 |
 | U.S. Global Ocean Ship-based Hydrographic Investigations Program (GO-SHIP) | https://usgoship.ucsd.edu/ |
 | World Ocean Atlas 2018 | https://www.ncei.noaa.gov/products/world-ocean-atlas |
 | Word Ocean Database | https://www.ncei.noaa.gov/products/world-ocean-database |
