### 2020-10-01 3d547fe
Update paper.md
