### 2020-10-01 6f4e3a5
Update paper.md
