+++ paper/paper.md
@@ -64,9 +64,10 @@ employ ``nctoolbox`` [@nctoolbox].
 | MOCHA Climatology | [doi:10.7282/T3XW4N4M](https://doi.org/10.7282/T3XW4N4M) |
 | HYbrid Coordinate Ocean Model | https://hycom.org |
 | CMEMS Global Ocean 1/12° Physics Analysis and Forecast | GLOBAL_ANALYSIS_FORECAST_ PHY_001_024 |
-| GO-SHIP | https://www.go-ship.org/ |
+| Global Ship-based Hydrographic
+Investigations Program (GO-SHIP) | https://www.go-ship.org/ |
 | World Ocean Atlas 2018 | https://www.ncei.noaa.gov/products/world-ocean-atlas |
-| Word Ocean Database | https://www.ncei.noaa.gov/products/world-ocean-database |
+| World Ocean Database | https://www.ncei.noaa.gov/products/world-ocean-database |
 
 : Data sources currently supported. \label{table:1}
 
