### 2020-10-01 fdd8cbc
Update paper.md
