### 2020-10-01 bc612d0
Update paper.md
