### 2020-10-01 lnferris/ocean_data_tools#36 @lnferris
I have added a table with this information to the README and paper
