### 2020-10-01 #2497 @lnferris
I have fixed and closed "DOIs of the datasets (lnferris/ocean_data_tools#36)".
