### 2020-10-05 #2497 @kthyng
At present, we need to hear from both reviewers @kakearney and @castelao regarding the reviewer checklists above which have not been completed to see what issues remain. We have already heard from @kakearney about a plan to check some tests, but that it will take a little bit to get back to this.
