### 2020-10-19 #2497 @kakearney
I believe I should have time in a few days; I will aim to have my re-review
completed by Oct. 26.

On Mon, Oct 19, 2020 at 2:03 PM Kristen Thyng <notifications@github.com>
wrote:

> Regarding the highlighted issues, lnferris/ocean_data_tools#28
> <https://github.com/lnferris/ocean_data_tools/issues/28> was only a
> suggestion on my part. lnferris/ocean_data_tools#30
> <https://github.com/lnferris/ocean_data_tools/issues/30> was more of a
> requirement; it looks like the recent updates do add much better input
> parsing and validation, but I have not yet had a chance to run any tests of
> my own (and likely won't for another two weeks due to workload).
>
> @kakearney <https://github.com/kakearney> Will you be able to return to
> this and run tests on the recent updates?
>
