### 2020-10-22 #2497 @kakearney
After perusing the updates, I have gone ahead and checked off the remaining items on my reviewer checklist.  Given the size of the toolbox and my current schedule, I wasn't able to run comprehensive tests on all the functions.  However, I did repeat my earlier running of the demos.m script, along with a few modifications to test input flexibility, and only encountered a few minor issues (see below).  I was very glad to see the updated input checks and improved documentation that addressed the bulk of my earlier comments.

I believe that the toolbox is now at a sufficient level for publication.  Thanks!

Demos.m comments:

- Line 244 of the demo still contains a hard-coded path; this should be added to the relative paths at the beginning of the demos.m script.
- Many path names in the demos.m script, and possibly elsewhere, assume *nix/Mac path conventions (i.e. forward slash as a path separator).  This may cause things to crash on Windows systems.  Path names should be constructed using `fullfile`, `pathsep`, etc. to keep them platform-independent.
- The input parsing and checking is much improved!  But there are still a few places where things fall through the cracks, notably with file input.  It would be nice if the various functions verified that files exist before they try to open them.  At the very least, some checks should be added to the demos.m file to make sure the user-modified bathymetry_dir path points to a valid file, and to error out gracefully if it does not.  For example, if a user were to ignore the instructions to modify the path (and I guarantee you most of them will!), this is the error they will get:

```
Error using netcdf.open (line 52)
Could not open file
'/Users/lnferris/Documents/data/bathymetry/topo_20.1.nc'.

Error in bathymetry_extract (line 51)
nc = netcdf.open(bathymetry_dir, 'NOWRITE'); % open the
file as netcdf datasource.

Error in demos (line 58)
bathymetry_plot(bathymetry_extract(bathymetry_dir,bounding_region(argo)),'2Dcontour')
```
