### 2020-10-26 #2497 @lnferris
If he does not come back, do we start over again?
