### 2020-10-27 #2497 @castelao
@kthyng , after going through the review checklist, I believe that this submission satisfies all the requirements for JOSS and is ready for publication.

Congratulations on the great work, @lnferris  .
