### 2020-10-29 #2497 @lnferris
I created a new release (1.0.1) and switched the cited DOI to this release.
