### 2020-10-30 #2497 @kthyng
@lnferris Ok I found the Zenodo doi you made (10.5281/zenodo.4151538). This will be linked into your publication formally through the paper separately from your citation of it, too. Can you go to the Zenodo archive and update the title and author list to exactly match your JOSS paper?
