### 2020-10-30 #2497 @kthyng
@lnferris I have a small change here you can merge if you agree [#40](https://github.com/lnferris/ocean_data_tools/pull/40)
