### 2020-10-30 #2497 @kthyng
@openjournals/dev the XML title doesn't have underscores that should be present. How should we proceed? Thanks.
