### 2020-10-30 #2497 @arfon
> @openjournals/dev the XML title doesn't have underscores that should be present. How should we proceed? Thanks.

Let's proceed with the submission and I'll fix this after we've published.
