### 2020-10-30 #2497 @kthyng
Congrats on your publication @lnferris!! Many thanks to reviewers @kakearney and @castelao for sharing your time and expertise. We couldn't do this process without you!!

once the DOI resolves, I'll close this issue. And @arfon will be able to fix the XML file at this point.
