### 2020-10-30 #2497 @arfon
> Ok DOI resolves but I guess the issue should stay open until we hear from @arfon.

Should all be fixed up now.
