### 2020-10-30 #2497 @lnferris
Thank you, @kthyng @kakearney @castelao @chadagreene for sharing your time and wisdom. This has been an awesome experience.
